Module pallet_staking::benchmarking

Staking pallet benchmarking.

Macros

benchmarks

Construct pallet benchmarks for weighing dispatchables.

impl_benchmark_test_suite

This creates a test suite which runs the module’s benchmarks.

whitelist_account

Functions

account

Grab an account, seeded by a name and index.

add_slashing_spans

create_validator_with_nominators

whitelisted_caller

This caller account is automatically whitelisted for DB reads/writes by the benchmarking macro.