Function pallet_staking::dispatchables::bond

pub fn bond<T: Config>(
    controller: <<T as Config>::Lookup as StaticLookup>::Source,
    value: BalanceOf<T>,
    payee: RewardDestination<T::AccountId>
)

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

value must be more than the minimum_balance specified by T::Currency.

The dispatch origin for this call must be Signed by the stash account.

Emits Bonded.

Complexity

Independent of the arguments. Moderate complexity.

O(1).

Three extra DB entries.

NOTE: Two of the storage writes (Self::bonded, Self::payee) are never cleaned unless the origin falls below existential deposit and gets removed as dust.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the bond variant of the pallet Call enum. You should not attempt to call this function directly.