Function pallet_staking::dispatchables::chill_other

pub fn chill_other<T: Config>(controller: T::AccountId)

Declare a controller to stop participating as either a validator or nominator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed, but can be called by anyone.

If the caller is the same as the controller being targeted, then no further checks are enforced, and this function behaves just like chill.

If the caller is different than the controller being targeted, the following conditions must be met:

controller must belong to a nominator who has become non-decodable,

Or:

A ChillThreshold must be set and checked which defines how close to the max nominators or validators we must reach before users can start chilling one-another.

A MaxNominatorCount and MaxValidatorCount must be set which is used to determine how close we are to the threshold.

A MinNominatorBond and MinValidatorBond must be set and checked, which determines if this is a person that should be chilled because they have not met the threshold bond required.

This can be helpful if bond requirements are updated, and we need to remove old users who do not satisfy these requirements.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::chill_other.