Function pallet_staking::dispatchables::reap_stash

pub fn reap_stash<T: Config>(stash: T::AccountId, num_slashing_spans: u32)

Remove all data structures concerning a staker/stash once it is at a state where it can be considered dust in the staking system. The requirements are:

the total_balance of the stash is below existential deposit.

or, the ledger.total of the stash is below existential deposit.

or, existential deposit is zero and either total_balance or ledger.total is zero.

The former can happen in cases like a slash; the latter when a fully unbonded account is still receiving staking rewards in RewardDestination::Staked.

It can be called by anyone, as long as stash meets the above requirements.

Refunds the transaction fees upon successful execution.

Parameters

num_slashing_spans: Refer to comments on Call::withdraw_unbonded for more details.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::reap_stash.