Function pallet_staking::dispatchables::set_staking_configs

pub fn set_staking_configs<T: Config>(
    min_nominator_bond: ConfigOp<BalanceOf<T>>,
    min_validator_bond: ConfigOp<BalanceOf<T>>,
    max_nominator_count: ConfigOp<u32>,
    max_validator_count: ConfigOp<u32>,
    chill_threshold: ConfigOp<Percent>,
    min_commission: ConfigOp<Perbill>,
    max_staked_rewards: ConfigOp<Percent>,
)

Update the various staking configurations .

min_nominator_bond: The minimum active bond needed to be a nominator.

min_validator_bond: The minimum active bond needed to be a validator.

max_nominator_count: The max number of users who can be a nominator at once. When set to None, no limit is enforced.

max_validator_count: The max number of users who can be a validator at once. When set to None, no limit is enforced.

chill_threshold: The ratio of max_nominator_count or max_validator_count which should be filled in order for the chill_other transaction to work.

min_commission: The minimum amount of commission that each validators must maintain. This is checked only upon calling validate. Existing validators are not affected.

RuntimeOrigin must be Root to call this function.

NOTE: Existing nominators and validators will not be affected by this update. to kick people under the new limits, chill_other should be called.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_staking_configs.