Enum pallet_staking::Call

pub enum Call<T: Config> {

    bond {
        controller: <T::Lookup as StaticLookup>::Source,
        value: BalanceOf<T>,
        payee: RewardDestination<T::AccountId>,
    },
    bond_extra {
        max_additional: BalanceOf<T>,
    },
    unbond {
        value: BalanceOf<T>,
    },
    withdraw_unbonded {
        num_slashing_spans: u32,
    },
    validate {
        prefs: ValidatorPrefs,
    },
    nominate {
        targets: Vec<<T::Lookup as StaticLookup>::Source>,
    },
    chill {},
    set_payee {
        payee: RewardDestination<T::AccountId>,
    },
    set_controller {
        controller: <T::Lookup as StaticLookup>::Source,
    },
    set_validator_count {
        new: u32,
    },
    increase_validator_count {
        additional: u32,
    },
    scale_validator_count {
        factor: Percent,
    },
    force_no_eras {},
    force_new_era {},
    set_invulnerables {
        invulnerables: Vec<T::AccountId>,
    },
    force_unstake {
        stash: T::AccountId,
        num_slashing_spans: u32,
    },
    force_new_era_always {},
    cancel_deferred_slash {
        era: EraIndex,
        slash_indices: Vec<u32>,
    },
    payout_stakers {
        validator_stash: T::AccountId,
        era: EraIndex,
    },
    rebond {
        value: BalanceOf<T>,
    },
    set_history_depth {
        new_history_depth: EraIndex,
        era_items_deleted: u32,
    },
    reap_stash {
        stash: T::AccountId,
        num_slashing_spans: u32,
    },
    kick {
        who: Vec<<T::Lookup as StaticLookup>::Source>,
    },
    set_staking_limits {
        min_nominator_bond: BalanceOf<T>,
        min_validator_bond: BalanceOf<T>,
        max_nominator_count: Option<u32>,
        max_validator_count: Option<u32>,
        threshold: Option<Percent>,
    },
    chill_other {
        controller: T::AccountId,
    },
    // some variants omitted

}

Contains one variant per dispatchable that can be called by an extrinsic.

Variants

bond

Fields

controller: <T::Lookup as StaticLookup>::Source

value: BalanceOf<T>

payee: RewardDestination<T::AccountId>

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

value must be more than the minimum_balance specified by T::Currency.

The dispatch origin for this call must be Signed by the stash account.

Emits Bonded.

Independent of the arguments. Moderate complexity.

O(1).

Three extra DB entries.

NOTE: Two of the storage writes (Self::bonded, Self::payee) are never cleaned unless the origin falls below existential deposit and gets removed as dust.

bond_extra

Fields

max_additional: BalanceOf<T>

Add some extra amount that have appeared in the stash free_balance into the balance up for staking.

The dispatch origin for this call must be Signed by the stash, not the controller.

Use this if there are additional funds in your stash account that you wish to bond. Unlike bond or unbond this function does not impose any limitation on the amount that can be added.

Emits Bonded.

Independent of the arguments. Insignificant complexity.

O(1).

unbond

Fields

value: BalanceOf<T>

Schedule a portion of the stash to be unlocked ready for transfer out after the bond period ends. If this leaves an amount actively bonded less than T::Currency::minimum_balance(), then it is increased to the full amount.

The dispatch origin for this call must be Signed by the controller, not the stash.

Once the unlock period is done, you can call withdraw_unbonded to actually move the funds out of management ready for transfer.

No more than a limited number of unlocking chunks (see MAX_UNLOCKING_CHUNKS) can co-exists at the same time. In that case, Call::withdraw_unbonded need to be called first to remove some of the chunks (if possible).

If a user encounters the InsufficientBond error when calling this extrinsic, they should call chill first in order to free up their bonded funds.

Emits Unbonded.

See also Call::withdraw_unbonded.

withdraw_unbonded

Fields

num_slashing_spans: u32

Remove any unlocked chunks from the unlocking queue from our management.

This essentially frees up that balance to be used by the stash account to do whatever it wants.

The dispatch origin for this call must be Signed by the controller.

Emits Withdrawn.

See also Call::unbond.

Complexity O(S) where S is the number of slashing spans to remove NOTE: Weight annotation is the kill scenario, we refund otherwise.

validate

Fields

prefs: ValidatorPrefs

Declare the desire to validate for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

nominate

Fields

targets: Vec<<T::Lookup as StaticLookup>::Source>

Declare the desire to nominate targets for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

The transaction’s complexity is proportional to the size of targets (N) which is capped at CompactAssignments::LIMIT (MAX_NOMINATIONS).

Both the reads and writes follow a similar pattern.

chill

Fields

Declare no desire to either validate or nominate.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

Independent of the arguments. Insignificant complexity.

Contains one read.

Writes are limited to the origin account key.

set_payee

Fields

payee: RewardDestination<T::AccountId>

(Re-)set the payment target for a controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1)

DB Weight:

Read: Ledger

Write: Payee

set_controller

Fields

controller: <T::Lookup as StaticLookup>::Source

(Re-)set the controller of a stash.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the stash, not the controller.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1) DB Weight:

Read: Bonded, Ledger New Controller, Ledger Old Controller

Write: Bonded, Ledger New Controller, Ledger Old Controller

set_validator_count

Fields

new: u32

Sets the ideal number of validators.

The dispatch origin must be Root.

Weight: O(1) Write: Validator Count

increase_validator_count

Fields

additional: u32

Increments the ideal number of validators.

The dispatch origin must be Root.

Same as Self::set_validator_count.

scale_validator_count

Fields

factor: Percent

Scale up the ideal number of validators by a factor.

The dispatch origin must be Root.

Same as Self::set_validator_count.

force_no_eras

Fields

Force there to be no new eras indefinitely.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. Thus the election process may be ongoing when this is called. In this case the election will continue until the next era is triggered.

No arguments.

Weight: O(1)

Write: ForceEra

force_new_era

Fields

Force there to be a new era at the end of the next session. After this, it will be reset to normal (non-forced) behaviour.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. If this is called just before a new era is triggered, the election process may not have enough blocks to get a result.

No arguments.

Weight: O(1)

Write ForceEra

set_invulnerables

Fields

invulnerables: Vec<T::AccountId>

Set the validators who cannot be slashed (if any).

The dispatch origin must be Root.

O(V)

Write: Invulnerables

force_unstake

Fields

stash: T::AccountId

num_slashing_spans: u32

Force a current staker to become completely unstaked, immediately.

The dispatch origin must be Root.

O(S) where S is the number of slashing spans to be removed Reads: Bonded, Slashing Spans, Account, Locks Writes: Bonded, Slashing Spans (if S > 0), Ledger, Payee, Validators, Nominators, Account, Locks Writes Each: SpanSlash * S

force_new_era_always

Fields

Force there to be a new era at the end of sessions indefinitely.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. If this is called just before a new era is triggered, the election process may not have enough blocks to get a result.

Weight: O(1)

Write: ForceEra

cancel_deferred_slash

Fields

era: EraIndex

slash_indices: Vec<u32>

Cancel enactment of a deferred slash.

Can be called by the T::SlashCancelOrigin.

Parameters: era and indices of the slashes for that era to kill.

Complexity: O(U + S) with U unapplied slashes weighted with U=1000 and S is the number of slash indices to be canceled.

Read: Unapplied Slashes

Write: Unapplied Slashes

payout_stakers

Fields

validator_stash: T::AccountId

era: EraIndex

Pay out all the stakers behind a single validator for a single era.

validator_stash is the stash account of the validator. Their nominators, up to T::MaxNominatorRewardedPerValidator, will also receive their rewards.

era may be any era between [current_era - history_depth; current_era].

The origin of this call must be Signed. Any account can call this function, even if it is not one of the stakers.

Time complexity: at most O(MaxNominatorRewardedPerValidator).

Contains a limited number of reads and writes.

N is the Number of payouts for the validator (including the validator) Weight:

Reward Destination Staked: O(N)

Reward Destination Controller (Creating): O(N)

NOTE: weights are assuming that payouts are made to alive stash account (Staked). Paying even a dead controller is cheaper weight-wise. We don’t do any refunds here.

rebond

Fields

value: BalanceOf<T>

Rebond a portion of the stash scheduled to be unlocked.

The dispatch origin must be signed by the controller.

Time complexity: O(L), where L is unlocking chunks

Bounded by MAX_UNLOCKING_CHUNKS.

Storage changes: Can’t increase storage, only decrease it.

set_history_depth

Fields

new_history_depth: EraIndex

era_items_deleted: u32

Set HistoryDepth value. This function will delete any history information when HistoryDepth is reduced.

Parameters:

new_history_depth: The new history depth you would like to set.

era_items_deleted: The number of items that will be deleted by this dispatch. This should report all the storage items that will be deleted by clearing old era history. Needed to report an accurate weight for the dispatch. Trusted by Root to report an accurate number.

Origin must be root.

E: Number of history depths removed, i.e. 10 -> 7 = 3

Weight: O(E)

DB Weight:

Reads: Current Era, History Depth

Writes: History Depth

Clear Prefix Each: Era Stakers, EraStakersClipped, ErasValidatorPrefs

Writes Each: ErasValidatorReward, ErasRewardPoints, ErasTotalStake, ErasStartSessionIndex

reap_stash

Fields

stash: T::AccountId

num_slashing_spans: u32

Remove all data structures concerning a staker/stash once it is at a state where it can be considered dust in the staking system. The requirements are:

the total_balance of the stash is below existential deposit.

or, the ledger.total of the stash is below existential deposit.

The former can happen in cases like a slash; the latter when a fully unbonded account is still receiving staking rewards in RewardDestination::Staked.

It can be called by anyone, as long as stash meets the above requirements.

Refunds the transaction fees upon successful execution.

kick

Fields

who: Vec<<T::Lookup as StaticLookup>::Source>

Remove the given nominations from the calling validator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

who: A list of nominator stash accounts who are nominating this validator which should no longer be nominating this validator.

Note: Making this call only makes sense if you first set the validator preferences to block any further nominations.

set_staking_limits

Fields

min_nominator_bond: BalanceOf<T>

min_validator_bond: BalanceOf<T>

max_nominator_count: Option<u32>

max_validator_count: Option<u32>

threshold: Option<Percent>

Update the various staking limits this pallet.

min_nominator_bond: The minimum active bond needed to be a nominator.

min_validator_bond: The minimum active bond needed to be a validator.

max_nominator_count: The max number of users who can be a nominator at once. When set to None, no limit is enforced.

max_validator_count: The max number of users who can be a validator at once. When set to None, no limit is enforced.

Origin must be Root to call this function.

NOTE: Existing nominators and validators will not be affected by this update. to kick people under the new limits, chill_other should be called.

chill_other

Fields

controller: T::AccountId

Declare a controller to stop participating as either a validator or nominator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed, but can be called by anyone.

If the caller is the same as the controller being targeted, then no further checks are enforced, and this function behaves just like chill.

If the caller is different than the controller being targeted, the following conditions must be met:

A ChillThreshold must be set and checked which defines how close to the max nominators or validators we must reach before users can start chilling one-another.

A MaxNominatorCount and MaxValidatorCount must be set which is used to determine how close we are to the threshold.

A MinNominatorBond and MinValidatorBond must be set and checked, which determines if this is a person that should be chilled because they have not met the threshold bond required.

This can be helpful if bond requirements are updated, and we need to remove old users who do not satisfy these requirements.

Implementations

impl<T: Config> Call<T>

pub fn new_call_variant_bond(
controller: <T::Lookup as StaticLookup>::Source,
value: BalanceOf<T>,
payee: RewardDestination<T::AccountId>
) -> Self

Create a call with the variant bond.

pub fn new_call_variant_bond_extra(max_additional: BalanceOf<T>) -> Self

Create a call with the variant bond_extra.

pub fn new_call_variant_unbond(value: BalanceOf<T>) -> Self

Create a call with the variant unbond.

pub fn new_call_variant_withdraw_unbonded(num_slashing_spans: u32) -> Self

Create a call with the variant withdraw_unbonded.

pub fn new_call_variant_validate(prefs: ValidatorPrefs) -> Self

Create a call with the variant validate.

pub fn new_call_variant_nominate(
targets: Vec<<T::Lookup as StaticLookup>::Source>
) -> Self

Create a call with the variant nominate.

pub fn new_call_variant_chill() -> Self

Create a call with the variant chill.

pub fn new_call_variant_set_payee(
payee: RewardDestination<T::AccountId>
) -> Self

Create a call with the variant set_payee.

pub fn new_call_variant_set_controller(
controller: <T::Lookup as StaticLookup>::Source
) -> Self

Create a call with the variant set_controller.

pub fn new_call_variant_set_validator_count(new: u32) -> Self

Create a call with the variant set_validator_count.

pub fn new_call_variant_increase_validator_count(additional: u32) -> Self

Create a call with the variant increase_validator_count.

pub fn new_call_variant_scale_validator_count(factor: Percent) -> Self

Create a call with the variant scale_validator_count.

pub fn new_call_variant_force_no_eras() -> Self

Create a call with the variant force_no_eras.

pub fn new_call_variant_force_new_era() -> Self

Create a call with the variant force_new_era.

pub fn new_call_variant_set_invulnerables(
invulnerables: Vec<T::AccountId>
) -> Self

Create a call with the variant set_invulnerables.

pub fn new_call_variant_force_unstake(
stash: T::AccountId,
num_slashing_spans: u32
) -> Self

Create a call with the variant force_unstake.

pub fn new_call_variant_force_new_era_always() -> Self

Create a call with the variant force_new_era_always.

pub fn new_call_variant_cancel_deferred_slash(
era: EraIndex,
slash_indices: Vec<u32>
) -> Self

Create a call with the variant cancel_deferred_slash.

pub fn new_call_variant_payout_stakers(
validator_stash: T::AccountId,
era: EraIndex
) -> Self

Create a call with the variant payout_stakers.

pub fn new_call_variant_rebond(value: BalanceOf<T>) -> Self

Create a call with the variant rebond.

pub fn new_call_variant_set_history_depth(
new_history_depth: EraIndex,
era_items_deleted: u32
) -> Self

Create a call with the variant set_history_depth.

pub fn new_call_variant_reap_stash(
stash: T::AccountId,
num_slashing_spans: u32
) -> Self

Create a call with the variant reap_stash.

pub fn new_call_variant_kick(
who: Vec<<T::Lookup as StaticLookup>::Source>
) -> Self

Create a call with the variant kick.

pub fn new_call_variant_set_staking_limits(
min_nominator_bond: BalanceOf<T>,
min_validator_bond: BalanceOf<T>,
max_nominator_count: Option<u32>,
max_validator_count: Option<u32>,
threshold: Option<Percent>
) -> Self

Create a call with the variant set_staking_limits.

pub fn new_call_variant_chill_other(controller: T::AccountId) -> Self

Create a call with the variant chill_other.