Struct pallet_staking::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn slashable_balance_of(stash: &T::AccountId) -> BalanceOf<T>

The total balance that can be slashed from a stash account as of right now.

pub fn slashable_balance_of_vote_weight(
stash: &T::AccountId,
issuance: BalanceOf<T>
) -> VoteWeight

Internal impl of Self::slashable_balance_of that returns VoteWeight.

pub fn weight_of_fn() -> Box<dyn Fn(&T::AccountId) -> VoteWeight>

Returns a closure around slashable_balance_of_vote_weight that can be passed around.

This prevents call sites from repeatedly requesting total_issuance from backend. But it is important to be only used while the total issuance is not changing.

pub fn weight_of(who: &T::AccountId) -> VoteWeight

Same as weight_of_fn, but made for one time use.

pub fn trigger_new_era(
start_session_index: SessionIndex,
exposures: Vec<(T::AccountId, Exposure<T::AccountId, BalanceOf<T>>)>
) -> Vec<T::AccountId>

Plan a new era.

Bump the current era storage (which holds the latest planned era).

Store start session index for the new planned era.

Clean old era information.

Store staking information for the new planned era

Returns the new validator set.

pub fn store_stakers_info(
exposures: Vec<(T::AccountId, Exposure<T::AccountId, BalanceOf<T>>)>,
new_planned_era: EraIndex
) -> Vec<T::AccountId>

Process the output of the election.

Store staking information for the new planned era

pub fn reward_by_ids(
validators_points: impl IntoIterator<Item = (T::AccountId, u32)>
)

Add reward points to validators using their stash account ID.

Validators are keyed by stash account ID and must be in the current elected set.

For each element in the iterator the given number of points in u32 is added to the validator, thus duplicates are handled.

At the end of the era each the total payout will be distributed among validator relatively to their points.

COMPLEXITY: Complexity is number_of_validator_to_reward x current_elected_len.

pub fn add_era_stakers(
current_era: EraIndex,
controller: T::AccountId,
exposure: Exposure<T::AccountId, BalanceOf<T>>
)

pub fn set_slash_reward_fraction(fraction: Perbill)

pub fn get_npos_voters(maybe_max_len: Option<usize>) -> Vec<VoterOf<Self>>

Get all of the voters that are eligible for the npos election.

maybe_max_len can imposes a cap on the number of voters returned; First all the validator are included in no particular order, then remainder is taken from the nominators, as returned by Config::SortedListProvider.

This will use nominators, and all the validators will inject a self vote.

This function is self-weighing as DispatchClass::Mandatory.

Slashing

All nominations that have been submitted before the last non-zero slash of the validator are auto-chilled, but still count towards the limit imposed by maybe_max_len.

pub fn get_npos_targets() -> Vec<T::AccountId>

Get the targets for an upcoming npos election.

This function is self-weighing as DispatchClass::Mandatory.

pub fn do_add_nominator(who: &T::AccountId, nominations: Nominations<T>)

This function will add a nominator to the Nominators storage map, and SortedListProvider.

If the nominator already exists, their nominations will be updated.

NOTE: you must ALWAYS use this function to add nominator or update their targets. Any access to Nominators or VoterList outside of this function is almost certainly wrong.

pub fn do_remove_nominator(who: &T::AccountId) -> bool

This function will remove a nominator from the Nominators storage map, and SortedListProvider.

Returns true if who was removed from Nominators, otherwise false.

NOTE: you must ALWAYS use this function to remove a nominator from the system. Any access to Nominators or VoterList outside of this function is almost certainly wrong.

pub fn do_add_validator(who: &T::AccountId, prefs: ValidatorPrefs)

This function will add a validator to the Validators storage map.

If the validator already exists, their preferences will be updated.

NOTE: you must ALWAYS use this function to add a validator to the system. Any access to Validators or VoterList outside of this function is almost certainly wrong.

pub fn do_remove_validator(who: &T::AccountId) -> bool

This function will remove a validator from the Validators storage map.

Returns true if who was removed from Validators, otherwise false.

NOTE: you must ALWAYS use this function to remove a validator from the system. Any access to Validators or VoterList outside of this function is almost certainly wrong.

impl<T: Config> Pallet<T>

pub fn bond(
origin: OriginFor<T>,
controller: <T::Lookup as StaticLookup>::Source,
value: BalanceOf<T>,
payee: RewardDestination<T::AccountId>
) -> DispatchResult

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

value must be more than the minimum_balance specified by T::Currency.

The dispatch origin for this call must be Signed by the stash account.

Emits Bonded.

Independent of the arguments. Moderate complexity.

O(1).

Three extra DB entries.

NOTE: Two of the storage writes (Self::bonded, Self::payee) are never cleaned unless the origin falls below existential deposit and gets removed as dust.

pub fn bond_extra(
origin: OriginFor<T>,
max_additional: BalanceOf<T>
) -> DispatchResult

Add some extra amount that have appeared in the stash free_balance into the balance up for staking.

The dispatch origin for this call must be Signed by the stash, not the controller.

Use this if there are additional funds in your stash account that you wish to bond. Unlike bond or unbond this function does not impose any limitation on the amount that can be added.

Emits Bonded.

Independent of the arguments. Insignificant complexity.

O(1).

pub fn unbond(origin: OriginFor<T>, value: BalanceOf<T>) -> DispatchResult

Schedule a portion of the stash to be unlocked ready for transfer out after the bond period ends. If this leaves an amount actively bonded less than T::Currency::minimum_balance(), then it is increased to the full amount.

The dispatch origin for this call must be Signed by the controller, not the stash.

Once the unlock period is done, you can call withdraw_unbonded to actually move the funds out of management ready for transfer.

No more than a limited number of unlocking chunks (see MAX_UNLOCKING_CHUNKS) can co-exists at the same time. In that case, Call::withdraw_unbonded need to be called first to remove some of the chunks (if possible).

If a user encounters the InsufficientBond error when calling this extrinsic, they should call chill first in order to free up their bonded funds.

Emits Unbonded.

See also Call::withdraw_unbonded.

pub fn withdraw_unbonded(
origin: OriginFor<T>,
num_slashing_spans: u32
) -> DispatchResultWithPostInfo

Remove any unlocked chunks from the unlocking queue from our management.

This essentially frees up that balance to be used by the stash account to do whatever it wants.

The dispatch origin for this call must be Signed by the controller.

Emits Withdrawn.

See also Call::unbond.

Complexity O(S) where S is the number of slashing spans to remove NOTE: Weight annotation is the kill scenario, we refund otherwise.

pub fn validate(origin: OriginFor<T>, prefs: ValidatorPrefs) -> DispatchResult

Declare the desire to validate for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

pub fn nominate(
origin: OriginFor<T>,
targets: Vec<<T::Lookup as StaticLookup>::Source>
) -> DispatchResult

Declare the desire to nominate targets for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

The transaction’s complexity is proportional to the size of targets (N) which is capped at CompactAssignments::LIMIT (T::MaxNominations).

Both the reads and writes follow a similar pattern.

pub fn chill(origin: OriginFor<T>) -> DispatchResult

Declare no desire to either validate or nominate.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

Independent of the arguments. Insignificant complexity.

Contains one read.

Writes are limited to the origin account key.

pub fn set_payee(
origin: OriginFor<T>,
payee: RewardDestination<T::AccountId>
) -> DispatchResult

(Re-)set the payment target for a controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1)

DB Weight:

Read: Ledger

Write: Payee

pub fn set_controller(
origin: OriginFor<T>,
controller: <T::Lookup as StaticLookup>::Source
) -> DispatchResult

(Re-)set the controller of a stash.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the stash, not the controller.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1) DB Weight:

Read: Bonded, Ledger New Controller, Ledger Old Controller

Write: Bonded, Ledger New Controller, Ledger Old Controller

pub fn set_validator_count(origin: OriginFor<T>, new: u32) -> DispatchResult

Sets the ideal number of validators.

The dispatch origin must be Root.

Weight: O(1) Write: Validator Count

pub fn increase_validator_count(
origin: OriginFor<T>,
additional: u32
) -> DispatchResult

Increments the ideal number of validators.

The dispatch origin must be Root.

Same as Self::set_validator_count.

pub fn scale_validator_count(
origin: OriginFor<T>,
factor: Percent
) -> DispatchResult

Scale up the ideal number of validators by a factor.

The dispatch origin must be Root.

Same as Self::set_validator_count.

pub fn force_no_eras(origin: OriginFor<T>) -> DispatchResult

Force there to be no new eras indefinitely.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. Thus the election process may be ongoing when this is called. In this case the election will continue until the next era is triggered.

No arguments.

Weight: O(1)

Write: ForceEra

pub fn force_new_era(origin: OriginFor<T>) -> DispatchResult

Force there to be a new era at the end of the next session. After this, it will be reset to normal (non-forced) behaviour.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. If this is called just before a new era is triggered, the election process may not have enough blocks to get a result.

No arguments.

Weight: O(1)

Write ForceEra

pub fn set_invulnerables(
origin: OriginFor<T>,
invulnerables: Vec<T::AccountId>
) -> DispatchResult

Set the validators who cannot be slashed (if any).

The dispatch origin must be Root.

pub fn force_unstake(
origin: OriginFor<T>,
stash: T::AccountId,
num_slashing_spans: u32
) -> DispatchResult

Force a current staker to become completely unstaked, immediately.

The dispatch origin must be Root.

pub fn force_new_era_always(origin: OriginFor<T>) -> DispatchResult

Force there to be a new era at the end of sessions indefinitely.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. If this is called just before a new era is triggered, the election process may not have enough blocks to get a result.

pub fn cancel_deferred_slash(
origin: OriginFor<T>,
era: EraIndex,
slash_indices: Vec<u32>
) -> DispatchResult

Cancel enactment of a deferred slash.

Can be called by the T::SlashCancelOrigin.

Parameters: era and indices of the slashes for that era to kill.

pub fn payout_stakers(
origin: OriginFor<T>,
validator_stash: T::AccountId,
era: EraIndex
) -> DispatchResultWithPostInfo

Pay out all the stakers behind a single validator for a single era.

validator_stash is the stash account of the validator. Their nominators, up to T::MaxNominatorRewardedPerValidator, will also receive their rewards.

era may be any era between [current_era - history_depth; current_era].

The origin of this call must be Signed. Any account can call this function, even if it is not one of the stakers.

Time complexity: at most O(MaxNominatorRewardedPerValidator).

Contains a limited number of reads and writes.

N is the Number of payouts for the validator (including the validator) Weight:

Reward Destination Staked: O(N)

Reward Destination Controller (Creating): O(N)

NOTE: weights are assuming that payouts are made to alive stash account (Staked). Paying even a dead controller is cheaper weight-wise. We don’t do any refunds here.

pub fn rebond(
origin: OriginFor<T>,
value: BalanceOf<T>
) -> DispatchResultWithPostInfo

Rebond a portion of the stash scheduled to be unlocked.

The dispatch origin must be signed by the controller.

Time complexity: O(L), where L is unlocking chunks

Bounded by MAX_UNLOCKING_CHUNKS.

Storage changes: Can’t increase storage, only decrease it.

pub fn set_history_depth(
origin: OriginFor<T>,
new_history_depth: EraIndex,
_era_items_deleted: u32
) -> DispatchResult

Set HistoryDepth value. This function will delete any history information when HistoryDepth is reduced.

Parameters:

new_history_depth: The new history depth you would like to set.

era_items_deleted: The number of items that will be deleted by this dispatch. This should report all the storage items that will be deleted by clearing old era history. Needed to report an accurate weight for the dispatch. Trusted by Root to report an accurate number.

Origin must be root.

E: Number of history depths removed, i.e. 10 -> 7 = 3

Weight: O(E)

DB Weight:

Reads: Current Era, History Depth

Writes: History Depth

Clear Prefix Each: Era Stakers, EraStakersClipped, ErasValidatorPrefs

Writes Each: ErasValidatorReward, ErasRewardPoints, ErasTotalStake, ErasStartSessionIndex

pub fn reap_stash(
origin: OriginFor<T>,
stash: T::AccountId,
num_slashing_spans: u32
) -> DispatchResultWithPostInfo

Remove all data structures concerning a staker/stash once it is at a state where it can be considered dust in the staking system. The requirements are:

the total_balance of the stash is below existential deposit.

or, the ledger.total of the stash is below existential deposit.

The former can happen in cases like a slash; the latter when a fully unbonded account is still receiving staking rewards in RewardDestination::Staked.

It can be called by anyone, as long as stash meets the above requirements.

Refunds the transaction fees upon successful execution.

pub fn kick(
origin: OriginFor<T>,
who: Vec<<T::Lookup as StaticLookup>::Source>
) -> DispatchResult

Remove the given nominations from the calling validator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

who: A list of nominator stash accounts who are nominating this validator which should no longer be nominating this validator.

Note: Making this call only makes sense if you first set the validator preferences to block any further nominations.

pub fn set_staking_configs(
origin: OriginFor<T>,
min_nominator_bond: BalanceOf<T>,
min_validator_bond: BalanceOf<T>,
max_nominator_count: Option<u32>,
max_validator_count: Option<u32>,
chill_threshold: Option<Percent>,
min_commission: Perbill
) -> DispatchResult

Update the various staking configurations .

min_nominator_bond: The minimum active bond needed to be a nominator.

min_validator_bond: The minimum active bond needed to be a validator.

max_nominator_count: The max number of users who can be a nominator at once. When set to None, no limit is enforced.

max_validator_count: The max number of users who can be a validator at once. When set to None, no limit is enforced.

chill_threshold: The ratio of max_nominator_count or max_validator_count which should be filled in order for the chill_other transaction to work.

min_commission: The minimum amount of commission that each validators must maintain. This is checked only upon calling validate. Existing validators are not affected.

Origin must be Root to call this function.

NOTE: Existing nominators and validators will not be affected by this update. to kick people under the new limits, chill_other should be called.

pub fn chill_other(
origin: OriginFor<T>,
controller: T::AccountId
) -> DispatchResult

Declare a controller to stop participating as either a validator or nominator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed, but can be called by anyone.

If the caller is the same as the controller being targeted, then no further checks are enforced, and this function behaves just like chill.

If the caller is different than the controller being targeted, the following conditions must be met:

controller must belong to a nominator who has become non-decodable,

Or:

A ChillThreshold must be set and checked which defines how close to the max nominators or validators we must reach before users can start chilling one-another.

A MaxNominatorCount and MaxValidatorCount must be set which is used to determine how close we are to the threshold.

A MinNominatorBond and MinValidatorBond must be set and checked, which determines if this is a person that should be chilled because they have not met the threshold bond required.

This can be helpful if bond requirements are updated, and we need to remove old users who do not satisfy these requirements.

pub fn force_apply_min_commission(
origin: OriginFor<T>,
validator_stash: T::AccountId
) -> DispatchResult

Force a validator to have at least the minimum commission. This will not affect a validator who already has a commission greater than or equal to the minimum. Any account can call this.

impl<T: Config> Pallet<T>

pub fn error_metadata() -> Option<PalletErrorMetadata>

impl<T: Config> Pallet<T>

pub fn history_depth() -> u32

Number of eras to keep in history.

Information is kept for eras in [current_era - history_depth; current_era].

Must be more than the number of eras delayed by session otherwise. I.e. active era must always be in history. I.e. active_era > current_era - history_depth must be guaranteed.

impl<T: Config> Pallet<T>

pub fn validator_count() -> u32

The ideal number of staking participants.

impl<T: Config> Pallet<T>

pub fn minimum_validator_count() -> u32

Minimum number of staking participants before emergency conditions are imposed.

impl<T: Config> Pallet<T>

pub fn invulnerables() -> Vec<T::AccountId>

Any validators that may never be slashed or forcibly kicked. It’s a Vec since they’re easy to initialize and the performance hit is minimal (we expect no more than four invulnerables) and restricted to testnets.

impl<T: Config> Pallet<T>

pub fn bonded<KArg>(k: KArg) -> Option<T::AccountId> where
KArg: EncodeLike<T::AccountId>,

Map from all locked “stash” accounts to the controller account.

impl<T: Config> Pallet<T>

pub fn ledger<KArg>(
k: KArg
) -> Option<StakingLedger<T::AccountId, BalanceOf<T>>> where
KArg: EncodeLike<T::AccountId>,

Map from all (unlocked) “controller” accounts to the info regarding the staking.

impl<T: Config> Pallet<T>

pub fn payee<KArg>(k: KArg) -> RewardDestination<T::AccountId> where
KArg: EncodeLike<T::AccountId>,

Where the reward payment should be made. Keyed by stash.

impl<T: Config> Pallet<T>

pub fn validators<KArg>(k: KArg) -> ValidatorPrefs where
KArg: EncodeLike<T::AccountId>,

The map from (wannabe) validator stash key to the preferences of that validator.

impl<T: Config> Pallet<T>

pub fn nominators<KArg>(k: KArg) -> Option<Nominations<T>> where
KArg: EncodeLike<T::AccountId>,

The map from nominator stash key to their nomination preferences, namely the validators that they wish to support.

Note that the keys of this storage map might become non-decodable in case the Config::MaxNominations configuration is decreased. In this rare case, these nominators are still existent in storage, their key is correct and retrievable (i.e. contains_key indicates that they exist), but their value cannot be decoded. Therefore, the non-decodable nominators will effectively not-exist, until they re-submit their preferences such that it is within the bounds of the newly set Config::MaxNominations.

This implies that ::iter_keys().count() and ::iter().count() might return different values for this map. Moreover, the main ::count() is aligned with the former, namely the number of keys that exist.

Lastly, if any of the nominators become non-decodable, they can be chilled immediately via Call::chill_other dispatchable by anyone.

impl<T: Config> Pallet<T>

pub fn current_era() -> Option<EraIndex>

The current era index.

This is the latest planned era, depending on how the Session pallet queues the validator set, it might be active or not.

impl<T: Config> Pallet<T>

pub fn active_era() -> Option<ActiveEraInfo>

The active era information, it holds index and start.

The active era is the era being currently rewarded. Validator set of this era must be equal to SessionInterface::validators.

impl<T: Config> Pallet<T>

pub fn eras_start_session_index<KArg>(k: KArg) -> Option<SessionIndex> where
KArg: EncodeLike<EraIndex>,

The session index at which the era start for the last HISTORY_DEPTH eras.

Note: This tracks the starting session (i.e. session index when era start being active) for the eras in [CurrentEra - HISTORY_DEPTH, CurrentEra].

impl<T: Config> Pallet<T>

pub fn eras_stakers<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Exposure<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

Exposure of validator at era.

This is keyed first by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras. If stakers hasn’t been set or has been removed then empty exposure is returned.

impl<T: Config> Pallet<T>

pub fn eras_stakers_clipped<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Exposure<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

Clipped Exposure of validator at era.

This is similar to ErasStakers but number of nominators exposed is reduced to the T::MaxNominatorRewardedPerValidator biggest stakers. (Note: the field total and own of the exposure remains unchanged). This is used to limit the i/o cost for the nominator payout.

This is keyed fist by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras. If stakers hasn’t been set or has been removed then empty exposure is returned.

impl<T: Config> Pallet<T>

pub fn eras_validator_prefs<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> ValidatorPrefs where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

Similar to ErasStakers, this holds the preferences of validators.

This is keyed first by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras.

impl<T: Config> Pallet<T>

pub fn eras_validator_reward<KArg>(k: KArg) -> Option<BalanceOf<T>> where
KArg: EncodeLike<EraIndex>,

The total validator era payout for the last HISTORY_DEPTH eras.

Eras that haven’t finished yet or has been removed doesn’t have reward.

impl<T: Config> Pallet<T>

pub fn eras_reward_points<KArg>(k: KArg) -> EraRewardPoints<T::AccountId> where
KArg: EncodeLike<EraIndex>,

Rewards for the last HISTORY_DEPTH eras. If reward hasn’t been set or has been removed then 0 reward is returned.

impl<T: Config> Pallet<T>

pub fn eras_total_stake<KArg>(k: KArg) -> BalanceOf<T> where
KArg: EncodeLike<EraIndex>,

The total amount staked for the last HISTORY_DEPTH eras. If total hasn’t been set or has been removed then 0 stake is returned.

impl<T: Config> Pallet<T>

pub fn force_era() -> Forcing

Mode of era forcing.

impl<T: Config> Pallet<T>

pub fn slash_reward_fraction() -> Perbill

The percentage of the slash that is distributed to reporters.

The rest of the slashed value is handled by the Slash.

impl<T: Config> Pallet<T>

pub fn canceled_payout() -> BalanceOf<T>

The amount of currency given to reporters of a slash event which was canceled by extraordinary circumstances (e.g. governance).

impl<T: Config> Pallet<T>

pub fn current_planned_session() -> SessionIndex

The last planned session scheduled by the session pallet.

This is basically in sync with the call to pallet_session::SessionManager::new_session.

impl<T: Config> Pallet<T>

pub fn offending_validators() -> Vec<(u32, bool)>

Indices of validators that have offended in the active era and whether they are currently disabled.

This value should be a superset of disabled validators since not all offences lead to the validator being disabled (if there was no slash). This is needed to track the percentage of validators that have offended in the current era, ensuring a new era is forced if OffendingValidatorsThreshold is reached. The vec is always kept sorted so that we can find whether a given validator has previously offended using binary search. It gets cleared when the era ends.