Module pallet_staking::storage_types

Auto-generated docs-only module listing all defined storage types for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Structs

ActiveEra

The active era information, it holds index and start.

Bonded

Map from all locked “stash” accounts to the controller account.

BondedEras

A mapping from still-bonded eras to the first session index of that era.

CanceledSlashPayout

The amount of currency given to reporters of a slash event which was canceled by extraordinary circumstances (e.g. governance).

ChillThreshold

The threshold for when users can start calling chill_other for other validators / nominators. The threshold is compared to the actual number of validators / nominators (CountFor*) in the system compared to the configured max (Max*Count).

CurrentEra

The current era index.

CurrentPlannedSession

The last planned session scheduled by the session pallet.

ErasRewardPoints

Rewards for the last HISTORY_DEPTH eras. If reward hasn’t been set or has been removed then 0 reward is returned.

ErasStakers

Exposure of validator at era.

ErasStakersClipped

Clipped Exposure of validator at era.

ErasStartSessionIndex

The session index at which the era start for the last HISTORY_DEPTH eras.

ErasTotalStake

The total amount staked for the last HISTORY_DEPTH eras. If total hasn’t been set or has been removed then 0 stake is returned.

ErasValidatorPrefs

Similar to ErasStakers, this holds the preferences of validators.

ErasValidatorReward

The total validator era payout for the last HISTORY_DEPTH eras.

ForceEra

Mode of era forcing.

Invulnerables

Any validators that may never be slashed or forcibly kicked. It’s a Vec since they’re easy to initialize and the performance hit is minimal (we expect no more than four invulnerables) and restricted to testnets.

Ledger

Map from all (unlocked) “controller” accounts to the info regarding the staking.

MaxNominatorsCount

The maximum nominator count before we stop allowing new validators to join.

MaxValidatorsCount

The maximum validator count before we stop allowing new validators to join.

MinCommission

The minimum amount of commission that validators can set.

MinNominatorBond

The minimum active bond to become and maintain the role of a nominator.

MinValidatorBond

The minimum active bond to become and maintain the role of a validator.

MinimumActiveStake

The minimum active nominator stake of the last successful election.

MinimumValidatorCount

Minimum number of staking participants before emergency conditions are imposed.

NominatorSlashInEra

All slashing events on nominators, mapped by era to the highest slash value of the era.

Nominators

The map from nominator stash key to their nomination preferences, namely the validators that they wish to support.

OffendingValidators

Indices of validators that have offended in the active era and whether they are currently disabled.

Payee

Where the reward payment should be made. Keyed by stash.

SlashRewardFraction

The percentage of the slash that is distributed to reporters.

SlashingSpans

Slashing spans for stash accounts.

SpanSlash

Records information about the maximum slash of a stash within a slashing span, as well as how much reward has been paid out.

UnappliedSlashes

All unapplied slashes that are queued for later.

ValidatorCount

The ideal number of active validators.

ValidatorSlashInEra

All slashing events on validators, mapped by era to the highest slash proportion and slash value of the era.

Validators

The map from (wannabe) validator stash key to the preferences of that validator.