Struct pallet_staking::Module

pub struct Module<T: Trait>(_);

Methods

impl<T: Trait + 'static> Module<T>

pub fn history_depth() -> u32

Number of era to keep in history.

Information is kept for eras in `[current_era - history_depth; current_era]

Must be more than the number of era delayed by session otherwise. i.e. active era must always be in history. i.e. active_era > current_era - history_depth must be guaranteed.

pub fn validator_count() -> u32

The ideal number of staking participants.

pub fn minimum_validator_count() -> u32

Minimum number of staking participants before emergency conditions are imposed.

pub fn invulnerables() -> Vec<T::AccountId>

Any validators that may never be slashed or forcibly kicked. It's a Vec since they're easy to initialize and the performance hit is minimal (we expect no more than four invulnerables) and restricted to testnets.

pub fn bonded<K: EncodeLike<T::AccountId>>(key: K) -> Option<T::AccountId>

Map from all locked "stash" accounts to the controller account.

pub fn ledger<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<StakingLedger<T::AccountId, BalanceOf<T>>>

Map from all (unlocked) "controller" accounts to the info regarding the staking.

pub fn payee<K: EncodeLike<T::AccountId>>(key: K) -> RewardDestination

Where the reward payment should be made. Keyed by stash.

pub fn validators<K: EncodeLike<T::AccountId>>(key: K) -> ValidatorPrefs

The map from (wannabe) validator stash key to the preferences of that validator.

pub fn nominators<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<Nominations<T::AccountId>>

The map from nominator stash key to the set of stash keys of all validators to nominate.

pub fn current_era() -> Option<EraIndex>

The current era index.

This is the latest planned era, depending on how session module queues the validator set, it might be active or not.

pub fn active_era() -> Option<ActiveEraInfo>

The active era information, it holds index and start.

The active era is the era currently rewarded. Validator set of this era must be equal to SessionInterface::validators.

pub fn eras_start_session_index<K: EncodeLike<EraIndex>>(
key: K
) -> Option<SessionIndex>

The session index at which the era start for the last HISTORY_DEPTH eras

pub fn eras_stakers<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Exposure<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

pub fn eras_stakers_clipped<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Exposure<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

pub fn eras_validator_prefs<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> ValidatorPrefs where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

pub fn eras_validator_reward<K: EncodeLike<EraIndex>>(
key: K
) -> Option<BalanceOf<T>>

The total validator era payout for the last HISTORY_DEPTH eras.

Eras that haven't finished yet or has been removed doesn't have reward.

pub fn eras_reward_points<K: EncodeLike<EraIndex>>(
key: K
) -> EraRewardPoints<T::AccountId>

Rewards for the last HISTORY_DEPTH eras. If reward hasn't been set or has been removed then 0 reward is returned.

pub fn eras_total_stake<K: EncodeLike<EraIndex>>(key: K) -> BalanceOf<T>

The total amount staked for the last HISTORY_DEPTH eras. If total hasn't been set or has been removed then 0 stake is returned.

pub fn force_era() -> Forcing

Mode of era forcing.

pub fn slash_reward_fraction() -> Perbill

The percentage of the slash that is distributed to reporters.

The rest of the slashed value is handled by the Slash.

pub fn canceled_payout() -> BalanceOf<T>

The amount of currency given to reporters of a slash event which was canceled by extraordinary circumstances (e.g. governance).

pub fn snapshot_validators() -> Option<Vec<T::AccountId>>

Snapshot of validators at the beginning of the current election window. This should only have a value when EraElectionStatus == ElectionStatus::Open(_).

pub fn snapshot_nominators() -> Option<Vec<T::AccountId>>

Snapshot of nominators at the beginning of the current election window. This should only have a value when EraElectionStatus == ElectionStatus::Open(_).

pub fn queued_elected() -> Option<ElectionResult<T::AccountId, BalanceOf<T>>>

The next validator set. At the end of an era, if this is available (potentially from the result of an offchain worker), it is immediately used. Otherwise, the on-chain election is executed.

pub fn queued_score() -> Option<PhragmenScore>

The score of the current QueuedElected.

pub fn era_election_status() -> ElectionStatus<T::BlockNumber>

Flag to control the execution of the offchain election. When Open(_), we accept solutions to be submitted.

pub fn is_current_session_final() -> bool

True if the current planned session is final.

impl<T: Trait> Module<T>

Can also be called using Call.

pub fn submit_election_solution(
origin: T::Origin,
winners: Vec<ValidatorIndex>,
compact_assignments: CompactAssignments,
score: PhragmenScore,
era: EraIndex
) -> DispatchResult

Submit a phragmen result to the chain. If the solution:

is valid.

has a better score than a potentially existing solution on chain.

then, it will be put on chain.

A solution consists of two pieces of data:

winners: a flat vector of all the winners of the round.

assignments: the compact version of an assignment vector that encodes the edge weights.

Both of which may be computed using [phragmen], or any other algorithm.

Additionally, the submitter must provide:

The score that they claim their solution has.

Both validators and nominators will be represented by indices in the solution. The indices should respect the corresponding types (ValidatorIndex and NominatorIndex). Moreover, they should be valid when used to index into SnapshotValidators and SnapshotNominators. Any invalid index will cause the solution to be rejected. These two storage items are set during the election window and may be used to determine the indices.

A solution is valid if:

It is submitted when EraElectionStatus is Open.

Its claimed score is equal to the score computed on-chain.

Presents the correct number of winners.

All indexes must be value according to the snapshot vectors. All edge values must also be correct and should not overflow the granularity of the ratio type (i.e. 256 or billion).

For each edge, all targets are actually nominated by the voter.

Has correct self-votes.

A solutions score is consisted of 3 parameters:

min { support.total } for each support of a winner. This value should be maximized.

sum { support.total } for each support of a winner. This value should be minimized.

sum { support.total^2 } for each support of a winner. This value should be minimized (to ensure less variance)

E: number of edges. m: size of winner committee. n: number of nominators. d: edge degree (16 for now) v: number of on-chain validator candidates.

NOTE: given a solution which is reduced, we can enable a new check the ensure |E| < n + m. We don't do this yet, but our offchain worker code executes it nonetheless.

major steps (all done in check_and_replace_solution):

Storage: O(1) read ElectionStatus.

Storage: O(1) read PhragmenScore.

Storage: O(1) read ValidatorCount.

Storage: O(1) length read from SnapshotValidators.

Storage: O(v) reads of AccountId to fetch snapshot_validators.

Memory: O(m) iterations to map winner index to validator id.

Storage: O(n) reads AccountId to fetch snapshot_nominators.

Memory: O(n + m) reads to map index to AccountId for un-compact.

Storage: O(e) accountid reads from Nomination to read correct nominations.

Storage: O(e) calls into slashable_balance_of_extended to convert ratio to staked.

Memory: build_support_map. O(e).

Memory: evaluate_support: O(E).

Storage: O(e) writes to QueuedElected.

Storage: O(1) write to QueuedScore

The weight of this call is 1/10th of the blocks total weight.

pub fn submit_election_solution_unsigned(
origin: T::Origin,
winners: Vec<ValidatorIndex>,
compact_assignments: CompactAssignments,
score: PhragmenScore,
era: EraIndex
) -> DispatchResult

Unsigned version of submit_election_solution.

Note that this must pass the [ValidateUnsigned] check which only allows transactions from the local node to be included. In other words, only the block author can include a transaction in the block.

impl<T: Trait> Module<T>

pub fn slashable_balance_of(stash: &T::AccountId) -> BalanceOf<T>

The total balance that can be slashed from a stash account as of right now.

pub fn pre_dispatch_checks(
score: PhragmenScore,
era: EraIndex
) -> Result<(), Error<T>>

Basic and cheap checks that we perform in validate unsigned, and in the execution.

pub fn check_and_replace_solution(
winners: Vec<ValidatorIndex>,
compact_assignments: CompactAssignments,
compute: ElectionCompute,
claimed_score: PhragmenScore,
era: EraIndex
) -> Result<(), Error<T>>

Checks a given solution and if correct and improved, writes it on chain as the queued result of the next round. This may be called by both a signed and an unsigned transaction.

pub fn reward_by_ids(
validators_points: impl IntoIterator<Item = (T::AccountId, u32)>
)

Add reward points to validators using their stash account ID.

Validators are keyed by stash account ID and must be in the current elected set.

For each element in the iterator the given number of points in u32 is added to the validator, thus duplicates are handled.

At the end of the era each the total payout will be distributed among validator relatively to their points.

COMPLEXITY: Complexity is number_of_validator_to_reward x current_elected_len. If you need to reward lots of validator consider using reward_by_indices.