pallet_staking

Trait EraPayout

pub trait EraPayout<Balance> {
    // Required method
    fn era_payout(
        total_staked: Balance,
        total_issuance: Balance,
        era_duration_millis: u64,
    ) -> (Balance, Balance);
}

Handler for determining how much of a balance should be paid out on the current era.

Required Methods

fn era_payout( total_staked: Balance, total_issuance: Balance, era_duration_millis: u64, ) -> (Balance, Balance)

Determine the payout for this era.

Returns the amount to be paid to stakers in this era, as well as whatever else should be paid out (“the rest”).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.