Type Alias pallet_staking::ClaimedRewards

pub type ClaimedRewards<T: Config> = StorageDoubleMap<_GeneratedPrefixForStorageClaimedRewards<T>, Twox64Concat, EraIndex, Twox64Concat, T::AccountId, Vec<Page>, ValueQuery>;

History of claimed paged rewards by era and validator.

This is keyed by era and validator stash which maps to the set of page indexes which have been claimed.

It is removed after Config::HistoryDepth eras.

Storage type is StorageDoubleMap with key1 type EraIndex, key2 type T :: AccountId and value type Vec < Page >.

Aliased Type

struct ClaimedRewards<T: Config>(/* private fields */);