Type Alias pallet_staking::ErasStakersPaged

pub type ErasStakersPaged<T: Config> = StorageNMap<_GeneratedPrefixForStorageErasStakersPaged<T>, (NMapKey<Twox64Concat, EraIndex>, NMapKey<Twox64Concat, T::AccountId>, NMapKey<Twox64Concat, Page>), ExposurePage<T::AccountId, BalanceOf<T>>, OptionQuery>;

Paginated exposure of a validator at given era.

This is keyed first by the era index to allow bulk deletion, then stash account and finally the page. Should only be accessed through EraInfo.

This is cleared after Config::HistoryDepth eras.

Storage type is StorageNMap with keys type (EraIndex, T :: AccountId, Page) and value type ExposurePage < T :: AccountId, BalanceOf < T > >.

Aliased Type

struct ErasStakersPaged<T: Config>(/* private fields */);