pallet_staking

Type Alias SlashRewardFraction

pub type SlashRewardFraction<T> = StorageValue<_GeneratedPrefixForStorageSlashRewardFraction<T>, Perbill, ValueQuery>;

The percentage of the slash that is distributed to reporters.

The rest of the slashed value is handled by the Slash.

Storage type is StorageValue with value type Perbill.

Aliased Type

pub struct SlashRewardFraction<T>(/* private fields */);