Module pallet_staking::weight

Functions

weight_for_correct_submit_solution
Weight of submit_solution in case of a correct submission.

weight_for_submit_solution
All weight notes are pertaining to the case of a better solution, in which we execute the longest code path. Weight: 0 + (0.63 μs * v) + (0.36 μs * n) + (96.53 μs * a ) + (8 μs * w ) with: