Function pallet_staking::weight::weight_for_submit_solution

pub fn weight_for_submit_solution<T: Trait>(
    winners: &Vec<ValidatorIndex>, 
    compact: &CompactAssignments, 
    size: &ElectionSize
) -> Weight

All weight notes are pertaining to the case of a better solution, in which we execute the longest code path. Weight: 0 + (0.63 μs * v) + (0.36 μs * n) + (96.53 μs * a ) + (8 μs * w ) with:

v validators in snapshot validators,

n nominators in snapshot nominators,

a assignment in the submitted solution

w winners in the submitted solution

State reads: - Initial checks: - ElectionState, CurrentEra, QueuedScore - SnapshotValidators.len() + SnapShotNominators.len() - ValidatorCount - SnapshotValidators - SnapshotNominators - Iterate over nominators: - compact.len() * Nominators(who) - (non_self_vote_edges) * SlashingSpans - For assignment_ratio_to_staked: Basically read the staked value of each stash. - (winners.len() + compact.len()) * (Ledger + Bonded) - TotalIssuance (read a gzillion times potentially, but well it is cached.)

State writes:

QueuedElected, QueuedScore