Enum pallet_staking::Call

pub enum Call<T: Config> {

    bond(<T::Lookup as StaticLookup>::Source, BalanceOf<T>, RewardDestination<T::AccountId>),
    bond_extra(BalanceOf<T>),
    unbond(BalanceOf<T>),
    withdraw_unbonded(u32),
    validate(ValidatorPrefs),
    nominate(Vec<<T::Lookup as StaticLookup>::Source>),
    chill(),
    set_payee(RewardDestination<T::AccountId>),
    set_controller(<T::Lookup as StaticLookup>::Source),
    set_validator_count(u32),
    increase_validator_count(u32),
    scale_validator_count(Percent),
    force_no_eras(),
    force_new_era(),
    set_invulnerables(Vec<T::AccountId>),
    force_unstake(T::AccountId, u32),
    force_new_era_always(),
    cancel_deferred_slash(EraIndex, Vec<u32>),
    payout_stakers(T::AccountId, EraIndex),
    rebond(BalanceOf<T>),
    set_history_depth(EraIndex, u32),
    reap_stash(T::AccountId, u32),
    submit_election_solution(Vec<ValidatorIndex>, CompactAssignments, ElectionScore, EraIndex, ElectionSize),
    submit_election_solution_unsigned(Vec<ValidatorIndex>, CompactAssignments, ElectionScore, EraIndex, ElectionSize),
    kick(Vec<<T::Lookup as StaticLookup>::Source>),
    // some variants omitted

}

Dispatchable calls.

Each variant of this enum maps to a dispatchable function from the associated module.

Variants

bond(<T::Lookup as StaticLookup>::Source, BalanceOf<T>, RewardDestination<T::AccountId>)

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

value must be more than the minimum_balance specified by T::Currency.

The dispatch origin for this call must be Signed by the stash account.

Emits Bonded.

Independent of the arguments. Moderate complexity.

O(1).

Three extra DB entries.

NOTE: Two of the storage writes (Self::bonded, Self::payee) are never cleaned unless the origin falls below existential deposit and gets removed as dust.

Weight: O(1) DB Weight:

Read: Bonded, Ledger, [Origin Account], Current Era, History Depth, Locks

Write: Bonded, Payee, [Origin Account], Locks, Ledger

Tuple Fields of bond

0: <T::Lookup as StaticLookup>::Source1: BalanceOf<T>2: RewardDestination<T::AccountId>

bond_extra(BalanceOf<T>)

Add some extra amount that have appeared in the stash free_balance into the balance up for staking.

Use this if there are additional funds in your stash account that you wish to bond. Unlike [bond] or [unbond] this function does not impose any limitation on the amount that can be added.

The dispatch origin for this call must be Signed by the stash, not the controller and it can be only called when EraElectionStatus is Closed.

Emits Bonded.

Independent of the arguments. Insignificant complexity.

O(1).

One DB entry.

DB Weight:

Read: Era Election Status, Bonded, Ledger, [Origin Account], Locks

Write: [Origin Account], Locks, Ledger

Tuple Fields of bond_extra

0: BalanceOf<T>

unbond(BalanceOf<T>)

Schedule a portion of the stash to be unlocked ready for transfer out after the bond period ends. If this leaves an amount actively bonded less than T::Currency::minimum_balance(), then it is increased to the full amount.

Once the unlock period is done, you can call withdraw_unbonded to actually move the funds out of management ready for transfer.

No more than a limited number of unlocking chunks (see MAX_UNLOCKING_CHUNKS) can co-exists at the same time. In that case, Call::withdraw_unbonded need to be called first to remove some of the chunks (if possible).

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when EraElectionStatus is Closed.

Emits Unbonded.

See also Call::withdraw_unbonded.

Independent of the arguments. Limited but potentially exploitable complexity.

Contains a limited number of reads.

Each call (requires the remainder of the bonded balance to be above minimum_balance) will cause a new entry to be inserted into a vector (Ledger.unlocking) kept in storage. The only way to clean the aforementioned storage item is also user-controlled via withdraw_unbonded.

One DB entry.

Weight: O(1) DB Weight:

Read: EraElectionStatus, Ledger, CurrentEra, Locks, BalanceOf Stash,

Write: Locks, Ledger, BalanceOf Stash,

Tuple Fields of unbond

0: BalanceOf<T>

withdraw_unbonded(u32)

Remove any unlocked chunks from the unlocking queue from our management.

This essentially frees up that balance to be used by the stash account to do whatever it wants.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when EraElectionStatus is Closed.

Emits Withdrawn.

See also Call::unbond.

Could be dependent on the origin argument and how much unlocking chunks exist. It implies consolidate_unlocked which loops over Ledger.unlocking, which is indirectly user-controlled. See [unbond] for more detail.

Contains a limited number of reads, yet the size of which could be large based on ledger.

Writes are limited to the origin account key.

Complexity O(S) where S is the number of slashing spans to remove Update:

Reads: EraElectionStatus, Ledger, Current Era, Locks, [Origin Account]

Writes: [Origin Account], Locks, Ledger Kill:

Reads: EraElectionStatus, Ledger, Current Era, Bonded, Slashing Spans, [Origin Account], Locks, BalanceOf stash

Writes: Bonded, Slashing Spans (if S > 0), Ledger, Payee, Validators, Nominators, [Origin Account], Locks, BalanceOf stash.

Writes Each: SpanSlash * S NOTE: Weight annotation is the kill scenario, we refund otherwise.

Tuple Fields of withdraw_unbonded

0: u32

validate(ValidatorPrefs)

Declare the desire to validate for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when EraElectionStatus is Closed.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1) DB Weight:

Read: Era Election Status, Ledger

Write: Nominators, Validators

Tuple Fields of validate

0: ValidatorPrefs

nominate(Vec<<T::Lookup as StaticLookup>::Source>)

Declare the desire to nominate targets for the origin controller.

Effects will be felt at the beginning of the next era. This can only be called when EraElectionStatus is Closed.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when EraElectionStatus is Closed.

The transaction’s complexity is proportional to the size of targets (N) which is capped at CompactAssignments::LIMIT (MAX_NOMINATIONS).

Both the reads and writes follow a similar pattern.

Weight: O(N) where N is the number of targets DB Weight:

Reads: Era Election Status, Ledger, Current Era

Writes: Validators, Nominators

Tuple Fields of nominate

0: Vec<<T::Lookup as StaticLookup>::Source>

chill()

Declare no desire to either validate or nominate.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when EraElectionStatus is Closed.

Independent of the arguments. Insignificant complexity.

Contains one read.

Writes are limited to the origin account key.

Weight: O(1) DB Weight:

Read: EraElectionStatus, Ledger

Write: Validators, Nominators

Tuple Fields of chill

set_payee(RewardDestination<T::AccountId>)

(Re-)set the payment target for a controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1)

DB Weight:

Read: Ledger

Write: Payee

Tuple Fields of set_payee

0: RewardDestination<T::AccountId>

set_controller(<T::Lookup as StaticLookup>::Source)

(Re-)set the controller of a stash.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the stash, not the controller.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1) DB Weight:

Read: Bonded, Ledger New Controller, Ledger Old Controller

Write: Bonded, Ledger New Controller, Ledger Old Controller

Tuple Fields of set_controller

0: <T::Lookup as StaticLookup>::Source

set_validator_count(u32)

Sets the ideal number of validators.

The dispatch origin must be Root.

Weight: O(1) Write: Validator Count

Tuple Fields of set_validator_count

0: u32

increase_validator_count(u32)

Increments the ideal number of validators.

The dispatch origin must be Root.

Same as [set_validator_count].

Tuple Fields of increase_validator_count

0: u32

scale_validator_count(Percent)

Scale up the ideal number of validators by a factor.

The dispatch origin must be Root.

Same as [set_validator_count].

Tuple Fields of scale_validator_count

0: Percent

force_no_eras()

Force there to be no new eras indefinitely.

The dispatch origin must be Root.

No arguments.

Weight: O(1)

Write: ForceEra

Tuple Fields of force_no_eras

force_new_era()

Force there to be a new era at the end of the next session. After this, it will be reset to normal (non-forced) behaviour.

The dispatch origin must be Root.

No arguments.

Weight: O(1)

Write ForceEra

Tuple Fields of force_new_era

set_invulnerables(Vec<T::AccountId>)

Set the validators who cannot be slashed (if any).

The dispatch origin must be Root.

O(V)

Write: Invulnerables

Tuple Fields of set_invulnerables

0: Vec<T::AccountId>

force_unstake(T::AccountId, u32)

Force a current staker to become completely unstaked, immediately.

The dispatch origin must be Root.

O(S) where S is the number of slashing spans to be removed Reads: Bonded, Slashing Spans, Account, Locks Writes: Bonded, Slashing Spans (if S > 0), Ledger, Payee, Validators, Nominators, Account, Locks Writes Each: SpanSlash * S

Tuple Fields of force_unstake

0: T::AccountId1: u32

force_new_era_always()

Force there to be a new era at the end of sessions indefinitely.

The dispatch origin must be Root.

Weight: O(1)

Write: ForceEra

Tuple Fields of force_new_era_always

cancel_deferred_slash(EraIndex, Vec<u32>)

Cancel enactment of a deferred slash.

Can be called by the T::SlashCancelOrigin.

Parameters: era and indices of the slashes for that era to kill.

Complexity: O(U + S) with U unapplied slashes weighted with U=1000 and S is the number of slash indices to be canceled.

Read: Unapplied Slashes

Write: Unapplied Slashes

Tuple Fields of cancel_deferred_slash

0: EraIndex1: Vec<u32>

payout_stakers(T::AccountId, EraIndex)

Pay out all the stakers behind a single validator for a single era.

validator_stash is the stash account of the validator. Their nominators, up to T::MaxNominatorRewardedPerValidator, will also receive their rewards.

era may be any era between [current_era - history_depth; current_era].

The origin of this call must be Signed. Any account can call this function, even if it is not one of the stakers.

This can only be called when EraElectionStatus is Closed.

Time complexity: at most O(MaxNominatorRewardedPerValidator).

Contains a limited number of reads and writes.

N is the Number of payouts for the validator (including the validator) Weight:

Reward Destination Staked: O(N)

Reward Destination Controller (Creating): O(N) DB Weight:

Read: EraElectionStatus, CurrentEra, HistoryDepth, ErasValidatorReward, ErasStakersClipped, ErasRewardPoints, ErasValidatorPrefs (8 items)

Read Each: Bonded, Ledger, Payee, Locks, System Account (5 items)

Write Each: System Account, Locks, Ledger (3 items)

NOTE: weights are assuming that payouts are made to alive stash account (Staked). Paying even a dead controller is cheaper weight-wise. We don’t do any refunds here.

Tuple Fields of payout_stakers

0: T::AccountId1: EraIndex

rebond(BalanceOf<T>)

Rebond a portion of the stash scheduled to be unlocked.

The dispatch origin must be signed by the controller, and it can be only called when EraElectionStatus is Closed.

Time complexity: O(L), where L is unlocking chunks

Bounded by MAX_UNLOCKING_CHUNKS.

Storage changes: Can’t increase storage, only decrease it.

DB Weight:

Reads: EraElectionStatus, Ledger, Locks, [Origin Account]

Writes: [Origin Account], Locks, Ledger

Tuple Fields of rebond

0: BalanceOf<T>

set_history_depth(EraIndex, u32)

Set HistoryDepth value. This function will delete any history information when HistoryDepth is reduced.

Parameters:

new_history_depth: The new history depth you would like to set.

era_items_deleted: The number of items that will be deleted by this dispatch. This should report all the storage items that will be deleted by clearing old era history. Needed to report an accurate weight for the dispatch. Trusted by Root to report an accurate number.

Origin must be root.

E: Number of history depths removed, i.e. 10 -> 7 = 3

Weight: O(E)

DB Weight:

Reads: Current Era, History Depth

Writes: History Depth

Clear Prefix Each: Era Stakers, EraStakersClipped, ErasValidatorPrefs

Writes Each: ErasValidatorReward, ErasRewardPoints, ErasTotalStake, ErasStartSessionIndex

Tuple Fields of set_history_depth

0: EraIndex1: u32

reap_stash(T::AccountId, u32)

Remove all data structure concerning a staker/stash once its balance is at the minimum. This is essentially equivalent to withdraw_unbonded except it can be called by anyone and the target stash must have no funds left beyond the ED.

This can be called from any origin.

stash: The stash account to reap. Its balance must be zero.

Complexity: O(S) where S is the number of slashing spans on the account. DB Weight:

Reads: Stash Account, Bonded, Slashing Spans, Locks

Writes: Bonded, Slashing Spans (if S > 0), Ledger, Payee, Validators, Nominators, Stash Account, Locks

Writes Each: SpanSlash * S

Tuple Fields of reap_stash

0: T::AccountId1: u32

submit_election_solution(Vec<ValidatorIndex>, CompactAssignments, ElectionScore, EraIndex, ElectionSize)

Submit an election result to the chain. If the solution:

is valid.

has a better score than a potentially existing solution on chain.

then, it will be put on chain.

A solution consists of two pieces of data:

winners: a flat vector of all the winners of the round.

assignments: the compact version of an assignment vector that encodes the edge weights.

Both of which may be computed using phragmen, or any other algorithm.

Additionally, the submitter must provide:

The score that they claim their solution has.

Both validators and nominators will be represented by indices in the solution. The indices should respect the corresponding types (ValidatorIndex and NominatorIndex). Moreover, they should be valid when used to index into SnapshotValidators and SnapshotNominators. Any invalid index will cause the solution to be rejected. These two storage items are set during the election window and may be used to determine the indices.

A solution is valid if:

It is submitted when EraElectionStatus is Open.

Its claimed score is equal to the score computed on-chain.

Presents the correct number of winners.

All indexes must be value according to the snapshot vectors. All edge values must also be correct and should not overflow the granularity of the ratio type (i.e. 256 or billion).

For each edge, all targets are actually nominated by the voter.

Has correct self-votes.

A solutions score is consisted of 3 parameters:

min { support.total } for each support of a winner. This value should be maximized.

sum { support.total } for each support of a winner. This value should be minimized.

sum { support.total^2 } for each support of a winner. This value should be minimized (to ensure less variance)

The transaction is assumed to be the longest path, a better solution.

Initial solution is almost the same.

Worse solution is retraced in pre-dispatch-checks which sets its own weight.

Tuple Fields of submit_election_solution

0: Vec<ValidatorIndex>1: CompactAssignments2: ElectionScore3: EraIndex4: ElectionSize

submit_election_solution_unsigned(Vec<ValidatorIndex>, CompactAssignments, ElectionScore, EraIndex, ElectionSize)

Unsigned version of submit_election_solution.

Note that this must pass the [ValidateUnsigned] check which only allows transactions from the local node to be included. In other words, only the block author can include a transaction in the block.

See [submit_election_solution].

Tuple Fields of submit_election_solution_unsigned

0: Vec<ValidatorIndex>1: CompactAssignments2: ElectionScore3: EraIndex4: ElectionSize

kick(Vec<<T::Lookup as StaticLookup>::Source>)

Remove the given nominations from the calling validator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when EraElectionStatus is Closed. The controller account should represent a validator.

who: A list of nominator stash accounts who are nominating this validator which should no longer be nominating this validator.

Note: Making this call only makes sense if you first set the validator preferences to block any further nominations.

Tuple Fields of kick

0: Vec<<T::Lookup as StaticLookup>::Source>