Enum pallet_staking::Call

pub enum Call<T: Trait> {
    bond(<T::Lookup as StaticLookup>::Source, BalanceOf<T>, RewardDestination),
    bond_extra(BalanceOf<T>),
    unbond(BalanceOf<T>),
    withdraw_unbonded(u32),
    validate(ValidatorPrefs),
    nominate(Vec<<T::Lookup as StaticLookup>::Source>),
    chill(),
    set_payee(RewardDestination),
    set_controller(<T::Lookup as StaticLookup>::Source),
    set_validator_count(u32),
    force_no_eras(),
    force_new_era(),
    set_invulnerables(Vec<T::AccountId>),
    force_unstake(T::AccountId, u32),
    force_new_era_always(),
    cancel_deferred_slash(EraIndex, Vec<u32>),
    payout_nominator(EraIndex, Vec<(T::AccountId, u32)>),
    payout_validator(EraIndex),
    payout_stakers(T::AccountId, EraIndex),
    rebond(BalanceOf<T>),
    set_history_depth(EraIndex, u32),
    reap_stash(T::AccountId, u32),
    submit_election_solution(Vec<ValidatorIndex>, CompactAssignments, PhragmenScore, EraIndex, ElectionSize),
    submit_election_solution_unsigned(Vec<ValidatorIndex>, CompactAssignments, PhragmenScore, EraIndex, ElectionSize),
    // some variants omitted
}

Dispatchable calls.

Each variant of this enum maps to a dispatchable function from the associated module.

Variants

bond(<T::Lookup as StaticLookup>::Source, BalanceOf<T>, RewardDestination)

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

value must be more than the minimum_balance specified by T::Currency.

The dispatch origin for this call must be Signed by the stash account.

Emits Bonded.

Independent of the arguments. Moderate complexity.

O(1).

Three extra DB entries.

NOTE: Two of the storage writes (Self::bonded, Self::payee) are never cleaned unless the origin falls below existential deposit and gets removed as dust.

Base Weight: 67.87 µs DB Weight:

Read: Bonded, Ledger, [Origin Account], Current Era, History Depth, Locks

Write: Bonded, Payee, [Origin Account], Locks, Ledger

bond_extra(BalanceOf<T>)

Add some extra amount that have appeared in the stash free_balance into the balance up for staking.

Use this if there are additional funds in your stash account that you wish to bond. Unlike [bond] or [unbond] this function does not impose any limitation on the amount that can be added.

The dispatch origin for this call must be Signed by the stash, not the controller and it can be only called when [EraElectionStatus] is Closed.

Emits Bonded.

Independent of the arguments. Insignificant complexity.

O(1).

One DB entry.

Base Weight: 54.88 µs DB Weight:

Read: Era Election Status, Bonded, Ledger, [Origin Account], Locks

Write: [Origin Account], Locks, Ledger

unbond(BalanceOf<T>)

Schedule a portion of the stash to be unlocked ready for transfer out after the bond period ends. If this leaves an amount actively bonded less than T::Currency::minimum_balance(), then it is increased to the full amount.

Once the unlock period is done, you can call withdraw_unbonded to actually move the funds out of management ready for transfer.

No more than a limited number of unlocking chunks (see MAX_UNLOCKING_CHUNKS) can co-exists at the same time. In that case, [Call::withdraw_unbonded] need to be called first to remove some of the chunks (if possible).

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

Emits Unbonded.

See also [Call::withdraw_unbonded].

Independent of the arguments. Limited but potentially exploitable complexity.

Contains a limited number of reads.

Each call (requires the remainder of the bonded balance to be above minimum_balance) will cause a new entry to be inserted into a vector (Ledger.unlocking) kept in storage. The only way to clean the aforementioned storage item is also user-controlled via withdraw_unbonded.

One DB entry.

Base Weight: 50.34 µs DB Weight:

Read: Era Election Status, Ledger, Current Era, Locks, [Origin Account]

Write: [Origin Account], Locks, Ledger

withdraw_unbonded(u32)

Remove any unlocked chunks from the unlocking queue from our management.

This essentially frees up that balance to be used by the stash account to do whatever it wants.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

Emits Withdrawn.

See also [Call::unbond].

Could be dependent on the origin argument and how much unlocking chunks exist. It implies consolidate_unlocked which loops over Ledger.unlocking, which is indirectly user-controlled. See [unbond] for more detail.

Contains a limited number of reads, yet the size of which could be large based on ledger.

Writes are limited to the origin account key.

Complexity O(S) where S is the number of slashing spans to remove Base Weight: Update: 50.52 + .028 * S µs

Reads: EraElectionStatus, Ledger, Current Era, Locks, [Origin Account]

Writes: [Origin Account], Locks, Ledger Kill: 79.41 + 2.366 * S µs

Reads: EraElectionStatus, Ledger, Current Era, Bonded, Slashing Spans, [Origin Account], Locks

Writes: Bonded, Slashing Spans (if S > 0), Ledger, Payee, Validators, Nominators, [Origin Account], Locks

Writes Each: SpanSlash * S NOTE: Weight annotation is the kill scenario, we refund otherwise.

validate(ValidatorPrefs)

Declare the desire to validate for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Base Weight: 17.13 µs DB Weight:

Read: Era Election Status, Ledger

Write: Nominators, Validators

nominate(Vec<<T::Lookup as StaticLookup>::Source>)

Declare the desire to nominate targets for the origin controller.

Effects will be felt at the beginning of the next era. This can only be called when [EraElectionStatus] is Closed.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

The transaction's complexity is proportional to the size of targets (N) which is capped at CompactAssignments::LIMIT (MAX_NOMINATIONS).

Both the reads and writes follow a similar pattern.

Base Weight: 22.34 + .36 * N µs where N is the number of targets DB Weight:

Reads: Era Election Status, Ledger, Current Era

Writes: Validators, Nominators

chill()

Declare no desire to either validate or nominate.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

Independent of the arguments. Insignificant complexity.

Contains one read.

Writes are limited to the origin account key.

Base Weight: 16.53 µs DB Weight:

Read: EraElectionStatus, Ledger

Write: Validators, Nominators

set_payee(RewardDestination)

(Re-)set the payment target for a controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Base Weight: 11.33 µs

DB Weight:

Read: Ledger

Write: Payee

set_controller(<T::Lookup as StaticLookup>::Source)

(Re-)set the controller of a stash.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the stash, not the controller.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Base Weight: 25.22 µs DB Weight:

Read: Bonded, Ledger New Controller, Ledger Old Controller

Write: Bonded, Ledger New Controller, Ledger Old Controller

set_validator_count(u32)

Sets the ideal number of validators.

The dispatch origin must be Root.

Base Weight: 1.717 µs Write: Validator Count

force_no_eras()

Force there to be no new eras indefinitely.

The dispatch origin must be Root.

No arguments.

Base Weight: 1.857 µs

Write: ForceEra

force_new_era()

Force there to be a new era at the end of the next session. After this, it will be reset to normal (non-forced) behaviour.

The dispatch origin must be Root.

No arguments.

Base Weight: 1.959 µs

Write ForceEra

set_invulnerables(Vec<T::AccountId>)

Set the validators who cannot be slashed (if any).

The dispatch origin must be Root.

O(V)

Base Weight: 2.208 + .006 * V µs

Write: Invulnerables

force_unstake(T::AccountId, u32)

Force a current staker to become completely unstaked, immediately.

The dispatch origin must be Root.

O(S) where S is the number of slashing spans to be removed Base Weight: 53.07 + 2.365 * S µs Reads: Bonded, Slashing Spans, Account, Locks Writes: Bonded, Slashing Spans (if S > 0), Ledger, Payee, Validators, Nominators, Account, Locks Writes Each: SpanSlash * S

force_new_era_always()

Force there to be a new era at the end of sessions indefinitely.

The dispatch origin must be Root.

Base Weight: 2.05 µs

Write: ForceEra

cancel_deferred_slash(EraIndex, Vec<u32>)

Cancel enactment of a deferred slash.

Can be called by either the root origin or the T::SlashCancelOrigin.

Parameters: era and indices of the slashes for that era to kill.

Complexity: O(U + S) with U unapplied slashes weighted with U=1000 and S is the number of slash indices to be canceled.

Base: 5870 + 34.61 * S µs

Read: Unapplied Slashes

Write: Unapplied Slashes

payout_nominator(EraIndex, Vec<(T::AccountId, u32)>)

This extrinsic will be removed after MigrationEra + HistoryDepth has passed, giving opportunity for users to claim all rewards before moving to Simple Payouts. After this time, you should use payout_stakers instead.

Make one nominator's payout for one era.

who is the controller account of the nominator to pay out.

era may not be lower than one following the most recently paid era. If it is higher, then it indicates an instruction to skip the payout of all previous eras.

validators is the list of all validators that who had exposure to during era, alongside the index of who in the clipped exposure of the validator. I.e. each element is a tuple of (validator, index of who in clipped exposure of validator). If it is incomplete, then less than the full reward will be paid out. It must not exceed MAX_NOMINATIONS.

WARNING: once an era is payed for a validator such validator can't claim the payout of previous era.

WARNING: Incorrect arguments here can result in loss of payout. Be very careful.

Number of storage read of O(validators); validators is the argument of the call, and is bounded by MAX_NOMINATIONS.

Each storage read is O(N) size and decode complexity; N is the maximum nominations that can be given to a single validator.

Computation complexity: O(MAX_NOMINATIONS * logN); MAX_NOMINATIONS is the maximum number of validators that may be nominated by a single nominator, it is bounded only economically (all nominators are required to place a minimum stake).

payout_validator(EraIndex)

This extrinsic will be removed after MigrationEra + HistoryDepth has passed, giving opportunity for users to claim all rewards before moving to Simple Payouts. After this time, you should use payout_stakers instead.

Make one validator's payout for one era.

who is the controller account of the validator to pay out.

era may not be lower than one following the most recently paid era. If it is higher, then it indicates an instruction to skip the payout of all previous eras.

WARNING: once an era is payed for a validator such validator can't claim the payout of previous era.

WARNING: Incorrect arguments here can result in loss of payout. Be very careful.

Time complexity: O(1).

Contains a limited number of reads and writes.

payout_stakers(T::AccountId, EraIndex)

Pay out all the stakers behind a single validator for a single era.

validator_stash is the stash account of the validator. Their nominators, up to T::MaxNominatorRewardedPerValidator, will also receive their rewards.

era may be any era between [current_era - history_depth; current_era].

The origin of this call must be Signed. Any account can call this function, even if it is not one of the stakers.

This can only be called when [EraElectionStatus] is Closed.

Time complexity: at most O(MaxNominatorRewardedPerValidator).

Contains a limited number of reads and writes.

N is the Number of payouts for the validator (including the validator) Base Weight: 110 + 54.2 * N µs (Median Slopes) DB Weight:

Read: EraElectionStatus, CurrentEra, HistoryDepth, MigrateEra, ErasValidatorReward, ErasStakersClipped, ErasRewardPoints, ErasValidatorPrefs (8 items)

Read Each: Bonded, Ledger, Payee, Locks, System Account (5 items)

Write Each: System Account, Locks, Ledger (3 items)

rebond(BalanceOf<T>)

Rebond a portion of the stash scheduled to be unlocked.

The dispatch origin must be signed by the controller, and it can be only called when [EraElectionStatus] is Closed.

Time complexity: O(L), where L is unlocking chunks

Bounded by MAX_UNLOCKING_CHUNKS.

Storage changes: Can't increase storage, only decrease it.

Base Weight: 34.51 µs * .048 L µs

DB Weight:

Reads: EraElectionStatus, Ledger, Locks, [Origin Account]

Writes: [Origin Account], Locks, Ledger

set_history_depth(EraIndex, u32)

Set HistoryDepth value. This function will delete any history information when HistoryDepth is reduced.

Parameters:

new_history_depth: The new history depth you would like to set.

era_items_deleted: The number of items that will be deleted by this dispatch. This should report all the storage items that will be deleted by clearing old era history. Needed to report an accurate weight for the dispatch. Trusted by Root to report an accurate number.

Origin must be root.

E: Number of history depths removed, i.e. 10 -> 7 = 3

Base Weight: 29.13 * E µs

DB Weight:

Reads: Current Era, History Depth

Writes: History Depth

Clear Prefix Each: Era Stakers, EraStakersClipped, ErasValidatorPrefs

Writes Each: ErasValidatorReward, ErasRewardPoints, ErasTotalStake, ErasStartSessionIndex

reap_stash(T::AccountId, u32)

Remove all data structure concerning a staker/stash once its balance is zero. This is essentially equivalent to withdraw_unbonded except it can be called by anyone and the target stash must have no funds left.

This can be called from any origin.

stash: The stash account to reap. Its balance must be zero.

Complexity: O(S) where S is the number of slashing spans on the account. Base Weight: 75.94 + 2.396 * S µs DB Weight:

Reads: Stash Account, Bonded, Slashing Spans, Locks

Writes: Bonded, Slashing Spans (if S > 0), Ledger, Payee, Validators, Nominators, Stash Account, Locks

Writes Each: SpanSlash * S

submit_election_solution(Vec<ValidatorIndex>, CompactAssignments, PhragmenScore, EraIndex, ElectionSize)

Submit a phragmen result to the chain. If the solution:

is valid.

has a better score than a potentially existing solution on chain.

then, it will be put on chain.

A solution consists of two pieces of data:

winners: a flat vector of all the winners of the round.

assignments: the compact version of an assignment vector that encodes the edge weights.

Both of which may be computed using [phragmen], or any other algorithm.

Additionally, the submitter must provide:

The score that they claim their solution has.

Both validators and nominators will be represented by indices in the solution. The indices should respect the corresponding types ([ValidatorIndex] and [NominatorIndex]). Moreover, they should be valid when used to index into [SnapshotValidators] and [SnapshotNominators]. Any invalid index will cause the solution to be rejected. These two storage items are set during the election window and may be used to determine the indices.

A solution is valid if:

It is submitted when [EraElectionStatus] is Open.

Its claimed score is equal to the score computed on-chain.

Presents the correct number of winners.

All indexes must be value according to the snapshot vectors. All edge values must also be correct and should not overflow the granularity of the ratio type (i.e. 256 or billion).

For each edge, all targets are actually nominated by the voter.

Has correct self-votes.

A solutions score is consisted of 3 parameters:

min { support.total } for each support of a winner. This value should be maximized.

sum { support.total } for each support of a winner. This value should be minimized.

sum { support.total^2 } for each support of a winner. This value should be minimized (to ensure less variance)

See crate::weight module.

submit_election_solution_unsigned(Vec<ValidatorIndex>, CompactAssignments, PhragmenScore, EraIndex, ElectionSize)

Unsigned version of submit_election_solution.

Note that this must pass the [ValidateUnsigned] check which only allows transactions from the local node to be included. In other words, only the block author can include a transaction in the block.

See crate::weight module.