Enum pallet_staking::Error

pub enum Error<T: Config> {

    NotController,
    NotStash,
    AlreadyBonded,
    AlreadyPaired,
    EmptyTargets,
    DuplicateIndex,
    InvalidSlashIndex,
    InsufficientValue,
    NoMoreChunks,
    NoUnlockChunk,
    FundedTarget,
    InvalidEraToReward,
    InvalidNumberOfNominations,
    NotSortedAndUnique,
    AlreadyClaimed,
    OffchainElectionEarlySubmission,
    OffchainElectionWeakSubmission,
    SnapshotUnavailable,
    OffchainElectionBogusWinnerCount,
    OffchainElectionBogusWinner,
    OffchainElectionBogusCompact,
    OffchainElectionBogusNominator,
    OffchainElectionBogusNomination,
    OffchainElectionSlashedNomination,
    OffchainElectionBogusSelfVote,
    OffchainElectionBogusEdge,
    OffchainElectionBogusScore,
    OffchainElectionBogusElectionSize,
    CallNotAllowed,
    IncorrectHistoryDepth,
    IncorrectSlashingSpans,
    BadState,
    TooManyTargets,
    BadTarget,
    // some variants omitted

}

Error for the staking module.

Variants

NotController

Not a controller account.

NotStash

Not a stash account.

AlreadyBonded

Stash is already bonded.

AlreadyPaired

Controller is already paired.

EmptyTargets

Targets cannot be empty.

DuplicateIndex

Duplicate index.

InvalidSlashIndex

Slash record index out of bounds.

InsufficientValue

Can not bond with value less than minimum balance.

NoMoreChunks

Can not schedule more unlock chunks.

NoUnlockChunk

Can not rebond without unlocking chunks.

FundedTarget

Attempting to target a stash that still has funds.

InvalidEraToReward

Invalid era to reward.

InvalidNumberOfNominations

Invalid number of nominations.

NotSortedAndUnique

Items are not sorted and unique.

AlreadyClaimed

Rewards for this era have already been claimed for this validator.

OffchainElectionEarlySubmission

The submitted result is received out of the open window.

OffchainElectionWeakSubmission

The submitted result is not as good as the one stored on chain.

SnapshotUnavailable

The snapshot data of the current window is missing.

OffchainElectionBogusWinnerCount

Incorrect number of winners were presented.

OffchainElectionBogusWinner

One of the submitted winners is not an active candidate on chain (index is out of range in snapshot).

OffchainElectionBogusCompact

Error while building the assignment type from the compact. This can happen if an index is invalid, or if the weights overflow.

OffchainElectionBogusNominator

One of the submitted nominators is not an active nominator on chain.

OffchainElectionBogusNomination

One of the submitted nominators has an edge to which they have not voted on chain.

OffchainElectionSlashedNomination

One of the submitted nominators has an edge which is submitted before the last non-zero slash of the target.

OffchainElectionBogusSelfVote

A self vote must only be originated from a validator to ONLY themselves.

OffchainElectionBogusEdge

The submitted result has unknown edges that are not among the presented winners.

OffchainElectionBogusScore

The claimed score does not match with the one computed from the data.

OffchainElectionBogusElectionSize

The election size is invalid.

CallNotAllowed

The call is not allowed at the given time due to restrictions of election period.

IncorrectHistoryDepth

Incorrect previous history depth input provided.

IncorrectSlashingSpans

Incorrect number of slashing spans provided.

BadState

Internal state has become somehow corrupted and the operation cannot continue.

TooManyTargets

Too many nomination targets supplied.

BadTarget

A nomination target was supplied that was blocked or otherwise not a validator.