Struct pallet_staking::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn bond(
origin: OriginFor<T>,
controller: <T::Lookup as StaticLookup>::Source,
value: BalanceOf<T>,
payee: RewardDestination<T::AccountId>
) -> DispatchResult

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

value must be more than the minimum_balance specified by T::Currency.

The dispatch origin for this call must be Signed by the stash account.

Emits Bonded.

Independent of the arguments. Moderate complexity.

O(1).

Three extra DB entries.

NOTE: Two of the storage writes (Self::bonded, Self::payee) are never cleaned unless the origin falls below existential deposit and gets removed as dust.

Weight: O(1) DB Weight:

Read: Bonded, Ledger, [Origin Account], Current Era, History Depth, Locks

Write: Bonded, Payee, [Origin Account], Locks, Ledger

pub fn bond_extra(
origin: OriginFor<T>,
max_additional: BalanceOf<T>
) -> DispatchResult

Add some extra amount that have appeared in the stash free_balance into the balance up for staking.

Use this if there are additional funds in your stash account that you wish to bond. Unlike [bond] or [unbond] this function does not impose any limitation on the amount that can be added.

The dispatch origin for this call must be Signed by the stash, not the controller and it can be only called when [EraElectionStatus] is Closed.

Emits Bonded.

Independent of the arguments. Insignificant complexity.

O(1).

One DB entry.

DB Weight:

Read: Era Election Status, Bonded, Ledger, [Origin Account], Locks

Write: [Origin Account], Locks, Ledger

pub fn unbond(origin: OriginFor<T>, value: BalanceOf<T>) -> DispatchResult

Schedule a portion of the stash to be unlocked ready for transfer out after the bond period ends. If this leaves an amount actively bonded less than T::Currency::minimum_balance(), then it is increased to the full amount.

Once the unlock period is done, you can call withdraw_unbonded to actually move the funds out of management ready for transfer.

No more than a limited number of unlocking chunks (see MAX_UNLOCKING_CHUNKS) can co-exists at the same time. In that case, Call::withdraw_unbonded need to be called first to remove some of the chunks (if possible).

If a user encounters the InsufficientBond error when calling this extrinsic, they should call chill first in order to free up their bonded funds.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

Emits Unbonded.

See also Call::withdraw_unbonded.

Independent of the arguments. Limited but potentially exploitable complexity.

Contains a limited number of reads.

Each call (requires the remainder of the bonded balance to be above minimum_balance) will cause a new entry to be inserted into a vector (Ledger.unlocking) kept in storage. The only way to clean the aforementioned storage item is also user-controlled via withdraw_unbonded.

One DB entry.

Weight: O(1) DB Weight:

Read: EraElectionStatus, Ledger, CurrentEra, Locks, BalanceOf Stash,

Write: Locks, Ledger, BalanceOf Stash,

pub fn withdraw_unbonded(
origin: OriginFor<T>,
num_slashing_spans: u32
) -> DispatchResultWithPostInfo

Remove any unlocked chunks from the unlocking queue from our management.

This essentially frees up that balance to be used by the stash account to do whatever it wants.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

Emits Withdrawn.

See also Call::unbond.

Could be dependent on the origin argument and how much unlocking chunks exist. It implies consolidate_unlocked which loops over Ledger.unlocking, which is indirectly user-controlled. See [unbond] for more detail.

Contains a limited number of reads, yet the size of which could be large based on ledger.

Writes are limited to the origin account key.

Complexity O(S) where S is the number of slashing spans to remove Update:

Reads: EraElectionStatus, Ledger, Current Era, Locks, [Origin Account]

Writes: [Origin Account], Locks, Ledger Kill:

Reads: EraElectionStatus, Ledger, Current Era, Bonded, Slashing Spans, [Origin Account], Locks, BalanceOf stash

Writes: Bonded, Slashing Spans (if S > 0), Ledger, Payee, Validators, Nominators, [Origin Account], Locks, BalanceOf stash.

Writes Each: SpanSlash * S NOTE: Weight annotation is the kill scenario, we refund otherwise.

pub fn validate(origin: OriginFor<T>, prefs: ValidatorPrefs) -> DispatchResult

Declare the desire to validate for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1) DB Weight:

Read: Era Election Status, Ledger

Write: Nominators, Validators

pub fn nominate(
origin: OriginFor<T>,
targets: Vec<<T::Lookup as StaticLookup>::Source>
) -> DispatchResult

Declare the desire to nominate targets for the origin controller.

Effects will be felt at the beginning of the next era. This can only be called when [EraElectionStatus] is Closed.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

The transaction’s complexity is proportional to the size of targets (N) which is capped at CompactAssignments::LIMIT (MAX_NOMINATIONS).

Both the reads and writes follow a similar pattern.

Weight: O(N) where N is the number of targets DB Weight:

Reads: Era Election Status, Ledger, Current Era

Writes: Validators, Nominators

pub fn chill(origin: OriginFor<T>) -> DispatchResult

Declare no desire to either validate or nominate.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

Independent of the arguments. Insignificant complexity.

Contains one read.

Writes are limited to the origin account key.

Weight: O(1) DB Weight:

Read: EraElectionStatus, Ledger

Write: Validators, Nominators

pub fn set_payee(
origin: OriginFor<T>,
payee: RewardDestination<T::AccountId>
) -> DispatchResult

(Re-)set the payment target for a controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1)

DB Weight:

Read: Ledger

Write: Payee

pub fn set_controller(
origin: OriginFor<T>,
controller: <T::Lookup as StaticLookup>::Source
) -> DispatchResult

(Re-)set the controller of a stash.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the stash, not the controller.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1) DB Weight:

Read: Bonded, Ledger New Controller, Ledger Old Controller

Write: Bonded, Ledger New Controller, Ledger Old Controller

pub fn set_validator_count(origin: OriginFor<T>, new: u32) -> DispatchResult

Sets the ideal number of validators.

The dispatch origin must be Root.

Weight: O(1) Write: Validator Count

pub fn increase_validator_count(
origin: OriginFor<T>,
additional: u32
) -> DispatchResult

Increments the ideal number of validators.

The dispatch origin must be Root.

Same as [set_validator_count].

pub fn scale_validator_count(
origin: OriginFor<T>,
factor: Percent
) -> DispatchResult

Scale up the ideal number of validators by a factor.

The dispatch origin must be Root.

Same as [set_validator_count].

pub fn force_no_eras(origin: OriginFor<T>) -> DispatchResult

Force there to be no new eras indefinitely.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. Thus the election process may be ongoing when this is called. In this case the election will continue until the next era is triggered.

No arguments.

Weight: O(1)

Write: ForceEra

pub fn force_new_era(origin: OriginFor<T>) -> DispatchResult

Force there to be a new era at the end of the next session. After this, it will be reset to normal (non-forced) behaviour.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. If this is called just before a new era is triggered, the election process may not have enough blocks to get a result.

No arguments.

Weight: O(1)

Write ForceEra

pub fn set_invulnerables(
origin: OriginFor<T>,
invulnerables: Vec<T::AccountId>
) -> DispatchResult

Set the validators who cannot be slashed (if any).

The dispatch origin must be Root.

O(V)

Write: Invulnerables

pub fn force_unstake(
origin: OriginFor<T>,
stash: T::AccountId,
num_slashing_spans: u32
) -> DispatchResult

Force a current staker to become completely unstaked, immediately.

The dispatch origin must be Root.

O(S) where S is the number of slashing spans to be removed Reads: Bonded, Slashing Spans, Account, Locks Writes: Bonded, Slashing Spans (if S > 0), Ledger, Payee, Validators, Nominators, Account, Locks Writes Each: SpanSlash * S

pub fn force_new_era_always(origin: OriginFor<T>) -> DispatchResult

Force there to be a new era at the end of sessions indefinitely.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. If this is called just before a new era is triggered, the election process may not have enough blocks to get a result.

Weight: O(1)

Write: ForceEra

pub fn cancel_deferred_slash(
origin: OriginFor<T>,
era: EraIndex,
slash_indices: Vec<u32>
) -> DispatchResult

Cancel enactment of a deferred slash.

Can be called by the T::SlashCancelOrigin.

Parameters: era and indices of the slashes for that era to kill.

Complexity: O(U + S) with U unapplied slashes weighted with U=1000 and S is the number of slash indices to be canceled.

Read: Unapplied Slashes

Write: Unapplied Slashes

pub fn payout_stakers(
origin: OriginFor<T>,
validator_stash: T::AccountId,
era: EraIndex
) -> DispatchResultWithPostInfo

Pay out all the stakers behind a single validator for a single era.

validator_stash is the stash account of the validator. Their nominators, up to T::MaxNominatorRewardedPerValidator, will also receive their rewards.

era may be any era between [current_era - history_depth; current_era].

The origin of this call must be Signed. Any account can call this function, even if it is not one of the stakers.

This can only be called when [EraElectionStatus] is Closed.

Time complexity: at most O(MaxNominatorRewardedPerValidator).

Contains a limited number of reads and writes.

N is the Number of payouts for the validator (including the validator) Weight:

Reward Destination Staked: O(N)

Reward Destination Controller (Creating): O(N) DB Weight:

Read: EraElectionStatus, CurrentEra, HistoryDepth, ErasValidatorReward, ErasStakersClipped, ErasRewardPoints, ErasValidatorPrefs (8 items)

Read Each: Bonded, Ledger, Payee, Locks, System Account (5 items)

Write Each: System Account, Locks, Ledger (3 items)

NOTE: weights are assuming that payouts are made to alive stash account (Staked). Paying even a dead controller is cheaper weight-wise. We don’t do any refunds here.

pub fn rebond(
origin: OriginFor<T>,
value: BalanceOf<T>
) -> DispatchResultWithPostInfo

Rebond a portion of the stash scheduled to be unlocked.

The dispatch origin must be signed by the controller, and it can be only called when [EraElectionStatus] is Closed.

Time complexity: O(L), where L is unlocking chunks

Bounded by MAX_UNLOCKING_CHUNKS.

Storage changes: Can’t increase storage, only decrease it.

DB Weight:

Reads: EraElectionStatus, Ledger, Locks, [Origin Account]

Writes: [Origin Account], Locks, Ledger

pub fn set_history_depth(
origin: OriginFor<T>,
new_history_depth: EraIndex,
_era_items_deleted: u32
) -> DispatchResult

Set HistoryDepth value. This function will delete any history information when HistoryDepth is reduced.

Parameters:

new_history_depth: The new history depth you would like to set.

era_items_deleted: The number of items that will be deleted by this dispatch. This should report all the storage items that will be deleted by clearing old era history. Needed to report an accurate weight for the dispatch. Trusted by Root to report an accurate number.

Origin must be root.

E: Number of history depths removed, i.e. 10 -> 7 = 3

Weight: O(E)

DB Weight:

Reads: Current Era, History Depth

Writes: History Depth

Clear Prefix Each: Era Stakers, EraStakersClipped, ErasValidatorPrefs

Writes Each: ErasValidatorReward, ErasRewardPoints, ErasTotalStake, ErasStartSessionIndex

pub fn reap_stash(
_origin: OriginFor<T>,
stash: T::AccountId,
num_slashing_spans: u32
) -> DispatchResult

Remove all data structure concerning a staker/stash once its balance is at the minimum. This is essentially equivalent to withdraw_unbonded except it can be called by anyone and the target stash must have no funds left beyond the ED.

This can be called from any origin.

stash: The stash account to reap. Its balance must be zero.

Complexity: O(S) where S is the number of slashing spans on the account. DB Weight:

Reads: Stash Account, Bonded, Slashing Spans, Locks

Writes: Bonded, Slashing Spans (if S > 0), Ledger, Payee, Validators, Nominators, Stash Account, Locks

Writes Each: SpanSlash * S

pub fn kick(
origin: OriginFor<T>,
who: Vec<<T::Lookup as StaticLookup>::Source>
) -> DispatchResult

Remove the given nominations from the calling validator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed. The controller account should represent a validator.

who: A list of nominator stash accounts who are nominating this validator which should no longer be nominating this validator.

Note: Making this call only makes sense if you first set the validator preferences to block any further nominations.

pub fn set_staking_limits(
origin: OriginFor<T>,
min_nominator_bond: BalanceOf<T>,
min_validator_bond: BalanceOf<T>,
max_nominator_count: Option<u32>,
max_validator_count: Option<u32>,
threshold: Option<Percent>
) -> DispatchResult

Update the various staking limits this pallet.

min_nominator_bond: The minimum active bond needed to be a nominator.

min_validator_bond: The minimum active bond needed to be a validator.

max_nominator_count: The max number of users who can be a nominator at once. When set to None, no limit is enforced.

max_validator_count: The max number of users who can be a validator at once. When set to None, no limit is enforced.

Origin must be Root to call this function.

NOTE: Existing nominators and validators will not be affected by this update. to kick people under the new limits, chill_other should be called.

pub fn chill_other(
origin: OriginFor<T>,
controller: T::AccountId
) -> DispatchResult

Declare a controller to stop participating as either a validator or nominator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed, but can be called by anyone.

If the caller is the same as the controller being targeted, then no further checks are enforced, and this function behaves just like chill.

If the caller is different than the controller being targeted, the following conditions must be met:

A ChillThreshold must be set and checked which defines how close to the max nominators or validators we must reach before users can start chilling one-another.

A MaxNominatorCount and MaxValidatorCount must be set which is used to determine how close we are to the threshold.

A MinNominatorBond and MinValidatorBond must be set and checked, which determines if this is a person that should be chilled because they have not met the threshold bond required.

This can be helpful if bond requirements are updated, and we need to remove old users who do not satisfy these requirements.

impl<T: Config> Pallet<T>

pub fn history_depth() -> u32

Number of eras to keep in history.

Information is kept for eras in [current_era - history_depth; current_era].

Must be more than the number of eras delayed by session otherwise. I.e. active era must always be in history. I.e. active_era > current_era - history_depth must be guaranteed.

impl<T: Config> Pallet<T>

pub fn validator_count() -> u32

The ideal number of staking participants.

impl<T: Config> Pallet<T>

pub fn minimum_validator_count() -> u32

Minimum number of staking participants before emergency conditions are imposed.

impl<T: Config> Pallet<T>

pub fn invulnerables() -> Vec<T::AccountId>

Any validators that may never be slashed or forcibly kicked. It’s a Vec since they’re easy to initialize and the performance hit is minimal (we expect no more than four invulnerables) and restricted to testnets.

impl<T: Config> Pallet<T>

pub fn bonded<KArg>(k: KArg) -> Option<T::AccountId> where
KArg: EncodeLike<T::AccountId>,

Map from all locked “stash” accounts to the controller account.

impl<T: Config> Pallet<T>

pub fn ledger<KArg>(
k: KArg
) -> Option<StakingLedger<T::AccountId, BalanceOf<T>>> where
KArg: EncodeLike<T::AccountId>,

Map from all (unlocked) “controller” accounts to the info regarding the staking.

impl<T: Config> Pallet<T>

pub fn payee<KArg>(k: KArg) -> RewardDestination<T::AccountId> where
KArg: EncodeLike<T::AccountId>,

Where the reward payment should be made. Keyed by stash.

impl<T: Config> Pallet<T>

pub fn validators<KArg>(k: KArg) -> ValidatorPrefs where
KArg: EncodeLike<T::AccountId>,

The map from (wannabe) validator stash key to the preferences of that validator.

When updating this storage item, you must also update the CounterForValidators.

impl<T: Config> Pallet<T>

pub fn nominators<KArg>(k: KArg) -> Option<Nominations<T::AccountId>> where
KArg: EncodeLike<T::AccountId>,

The map from nominator stash key to the set of stash keys of all validators to nominate.

When updating this storage item, you must also update the CounterForNominators.

impl<T: Config> Pallet<T>

pub fn current_era() -> Option<EraIndex>

The current era index.

This is the latest planned era, depending on how the Session pallet queues the validator set, it might be active or not.

impl<T: Config> Pallet<T>

pub fn active_era() -> Option<ActiveEraInfo>

The active era information, it holds index and start.

The active era is the era being currently rewarded. Validator set of this era must be equal to SessionInterface::validators.

impl<T: Config> Pallet<T>

pub fn eras_start_session_index<KArg>(k: KArg) -> Option<SessionIndex> where
KArg: EncodeLike<EraIndex>,

The session index at which the era start for the last HISTORY_DEPTH eras.

Note: This tracks the starting session (i.e. session index when era start being active) for the eras in [CurrentEra - HISTORY_DEPTH, CurrentEra].

impl<T: Config> Pallet<T>

pub fn eras_stakers<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Exposure<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

Exposure of validator at era.

This is keyed first by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras. If stakers hasn’t been set or has been removed then empty exposure is returned.

impl<T: Config> Pallet<T>

pub fn eras_stakers_clipped<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Exposure<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

Clipped Exposure of validator at era.

This is similar to ErasStakers but number of nominators exposed is reduced to the T::MaxNominatorRewardedPerValidator biggest stakers. (Note: the field total and own of the exposure remains unchanged). This is used to limit the i/o cost for the nominator payout.

This is keyed fist by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras. If stakers hasn’t been set or has been removed then empty exposure is returned.

impl<T: Config> Pallet<T>

pub fn eras_validator_prefs<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> ValidatorPrefs where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

Similar to ErasStakers, this holds the preferences of validators.

This is keyed first by the era index to allow bulk deletion and then the stash account.

Is it removed after HISTORY_DEPTH eras.

impl<T: Config> Pallet<T>

pub fn eras_validator_reward<KArg>(k: KArg) -> Option<BalanceOf<T>> where
KArg: EncodeLike<EraIndex>,

The total validator era payout for the last HISTORY_DEPTH eras.

Eras that haven’t finished yet or has been removed doesn’t have reward.

impl<T: Config> Pallet<T>

pub fn eras_reward_points<KArg>(k: KArg) -> EraRewardPoints<T::AccountId> where
KArg: EncodeLike<EraIndex>,

Rewards for the last HISTORY_DEPTH eras. If reward hasn’t been set or has been removed then 0 reward is returned.

impl<T: Config> Pallet<T>

pub fn eras_total_stake<KArg>(k: KArg) -> BalanceOf<T> where
KArg: EncodeLike<EraIndex>,

The total amount staked for the last HISTORY_DEPTH eras. If total hasn’t been set or has been removed then 0 stake is returned.

impl<T: Config> Pallet<T>

pub fn force_era() -> Forcing

Mode of era forcing.

impl<T: Config> Pallet<T>

pub fn slash_reward_fraction() -> Perbill

The percentage of the slash that is distributed to reporters.

The rest of the slashed value is handled by the Slash.

impl<T: Config> Pallet<T>

pub fn canceled_payout() -> BalanceOf<T>

The amount of currency given to reporters of a slash event which was canceled by extraordinary circumstances (e.g. governance).

impl<T: Config> Pallet<T>

pub fn current_planned_session() -> SessionIndex

The last planned session scheduled by the session pallet.

This is basically in sync with the call to [SessionManager::new_session].

impl<T: Config> Pallet<T>

pub fn slashable_balance_of(stash: &T::AccountId) -> BalanceOf<T>

The total balance that can be slashed from a stash account as of right now.

pub fn slashable_balance_of_vote_weight(
stash: &T::AccountId,
issuance: BalanceOf<T>
) -> VoteWeight

Internal impl of Self::slashable_balance_of that returns VoteWeight.

pub fn slashable_balance_of_fn() -> Box<dyn Fn(&T::AccountId) -> VoteWeight>

Returns a closure around slashable_balance_of_vote_weight that can be passed around.

This prevents call sites from repeatedly requesting total_issuance from backend. But it is important to be only used while the total issuance is not changing.

pub fn trigger_new_era(
start_session_index: SessionIndex,
exposures: Vec<(T::AccountId, Exposure<T::AccountId, BalanceOf<T>>)>
) -> Vec<T::AccountId>

Plan a new era.

Bump the current era storage (which holds the latest planned era).

Store start session index for the new planned era.

Clean old era information.

Store staking information for the new planned era

Returns the new validator set.

pub fn store_stakers_info(
exposures: Vec<(T::AccountId, Exposure<T::AccountId, BalanceOf<T>>)>,
new_planned_era: EraIndex
) -> Vec<T::AccountId>

Process the output of the election.

Store staking information for the new planned era

pub fn reward_by_ids(
validators_points: impl IntoIterator<Item = (T::AccountId, u32)>
)

Add reward points to validators using their stash account ID.

Validators are keyed by stash account ID and must be in the current elected set.

For each element in the iterator the given number of points in u32 is added to the validator, thus duplicates are handled.

At the end of the era each the total payout will be distributed among validator relatively to their points.

COMPLEXITY: Complexity is number_of_validator_to_reward x current_elected_len.

pub fn get_npos_voters() -> Vec<(T::AccountId, VoteWeight, Vec<T::AccountId>)>

Get all of the voters that are eligible for the npos election.

This will use all on-chain nominators, and all the validators will inject a self vote.

Slashing

All nominations that have been submitted before the last non-zero slash of the validator are auto-chilled.

Note that this is VERY expensive. Use with care.

pub fn get_npos_targets() -> Vec<T::AccountId>

This is a very expensive function and result should be cached versus being called multiple times.

pub fn do_add_nominator(
who: &T::AccountId,
nominations: Nominations<T::AccountId>
)

This function will add a nominator to the Nominators storage map, and keep track of the CounterForNominators.

If the nominator already exists, their nominations will be updated.

pub fn do_remove_nominator(who: &T::AccountId)

This function will remove a nominator from the Nominators storage map, and keep track of the CounterForNominators.

pub fn do_add_validator(who: &T::AccountId, prefs: ValidatorPrefs)

This function will add a validator to the Validators storage map, and keep track of the CounterForValidators.

If the validator already exists, their preferences will be updated.

pub fn do_remove_validator(who: &T::AccountId)

This function will remove a validator from the Validators storage map, and keep track of the CounterForValidators.