Trait pallet_staking::pallet::SessionInterface

pub trait SessionInterface<AccountId>: Config {
    fn disable_validator(validator: &AccountId);
    fn validators() -> Vec<AccountId>;
    fn prune_historical_up_to(up_to: SessionIndex);
}

Means for interacting with a specialized version of the session trait.

This is needed because Staking sets the ValidatorIdOf of the pallet_session::Config

Required Methods

fn disable_validator(validator: &AccountId)

Disable a given validator by stash ID.

Returns true if new era should be forced at the end of this session. This allows preventing a situation where there is too many validators disabled and block production stalls.

fn validators() -> Vec<AccountId>

Get the validators from session.

fn prune_historical_up_to(up_to: SessionIndex)

Prune historical session tries up to but not including the given index.