Struct pallet_staking::Module

pub struct Module<T: Config>(_);

Implementations

impl<T: Config + 'static> Module<T>

pub fn history_depth() -> u32

Number of eras to keep in history.

Information is kept for eras in [current_era - history_depth; current_era].

Must be more than the number of eras delayed by session otherwise. I.e. active era must always be in history. I.e. active_era > current_era - history_depth must be guaranteed.

pub fn validator_count() -> u32

The ideal number of staking participants.

pub fn minimum_validator_count() -> u32

Minimum number of staking participants before emergency conditions are imposed.

pub fn invulnerables() -> Vec<T::AccountId>

Any validators that may never be slashed or forcibly kicked. It’s a Vec since they’re easy to initialize and the performance hit is minimal (we expect no more than four invulnerables) and restricted to testnets.

pub fn bonded<K: EncodeLike<T::AccountId>>(key: K) -> Option<T::AccountId>

Map from all locked “stash” accounts to the controller account.

pub fn ledger<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<StakingLedger<T::AccountId, BalanceOf<T>>>

Map from all (unlocked) “controller” accounts to the info regarding the staking.

pub fn payee<K: EncodeLike<T::AccountId>>(
key: K
) -> RewardDestination<T::AccountId>

Where the reward payment should be made. Keyed by stash.

pub fn validators<K: EncodeLike<T::AccountId>>(key: K) -> ValidatorPrefs

The map from (wannabe) validator stash key to the preferences of that validator.

pub fn nominators<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<Nominations<T::AccountId>>

The map from nominator stash key to the set of stash keys of all validators to nominate.

pub fn current_era() -> Option<EraIndex>

The current era index.

This is the latest planned era, depending on how the Session pallet queues the validator set, it might be active or not.

pub fn active_era() -> Option<ActiveEraInfo>

The active era information, it holds index and start.

The active era is the era being currently rewarded. Validator set of this era must be equal to SessionInterface::validators.

pub fn eras_start_session_index<K: EncodeLike<EraIndex>>(
key: K
) -> Option<SessionIndex>

The session index at which the era start for the last HISTORY_DEPTH eras.

Note: This tracks the starting session (i.e. session index when era start being active) for the eras in [CurrentEra - HISTORY_DEPTH, CurrentEra].

pub fn eras_stakers<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Exposure<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

pub fn eras_stakers_clipped<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Exposure<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

pub fn eras_validator_prefs<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> ValidatorPrefs where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

pub fn eras_validator_reward<K: EncodeLike<EraIndex>>(
key: K
) -> Option<BalanceOf<T>>

The total validator era payout for the last HISTORY_DEPTH eras.

Eras that haven’t finished yet or has been removed doesn’t have reward.

pub fn eras_reward_points<K: EncodeLike<EraIndex>>(
key: K
) -> EraRewardPoints<T::AccountId>

Rewards for the last HISTORY_DEPTH eras. If reward hasn’t been set or has been removed then 0 reward is returned.

pub fn eras_total_stake<K: EncodeLike<EraIndex>>(key: K) -> BalanceOf<T>

The total amount staked for the last HISTORY_DEPTH eras. If total hasn’t been set or has been removed then 0 stake is returned.

pub fn force_era() -> Forcing

Mode of era forcing.

pub fn slash_reward_fraction() -> Perbill

The percentage of the slash that is distributed to reporters.

The rest of the slashed value is handled by the Slash.

pub fn canceled_payout() -> BalanceOf<T>

The amount of currency given to reporters of a slash event which was canceled by extraordinary circumstances (e.g. governance).

pub fn slashing_spans<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<SlashingSpans>

Slashing spans for stash accounts.

pub fn snapshot_validators() -> Option<Vec<T::AccountId>>

Snapshot of validators at the beginning of the current election window. This should only have a value when EraElectionStatus == ElectionStatus::Open(_).

pub fn snapshot_nominators() -> Option<Vec<T::AccountId>>

Snapshot of nominators at the beginning of the current election window. This should only have a value when EraElectionStatus == ElectionStatus::Open(_).

pub fn queued_elected() -> Option<ElectionResult<T::AccountId, BalanceOf<T>>>

The next validator set. At the end of an era, if this is available (potentially from the result of an offchain worker), it is immediately used. Otherwise, the on-chain election is executed.

pub fn queued_score() -> Option<ElectionScore>

The score of the current QueuedElected.

pub fn era_election_status() -> ElectionStatus<T::BlockNumber>

Flag to control the execution of the offchain election. When Open(_), we accept solutions to be submitted.

pub fn is_current_session_final() -> bool

True if the current planned session is final. Note that this does not take era forcing into account.

impl<T: Config> Module<T>

Can also be called using Call.

pub fn bond(
origin: T::Origin,
controller: <T::Lookup as StaticLookup>::Source,
value: BalanceOf<T>,
payee: RewardDestination<T::AccountId>
) -> DispatchResult

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

value must be more than the minimum_balance specified by T::Currency.

The dispatch origin for this call must be Signed by the stash account.

Emits Bonded.

Independent of the arguments. Moderate complexity.

O(1).

Three extra DB entries.

NOTE: Two of the storage writes (Self::bonded, Self::payee) are never cleaned unless the origin falls below existential deposit and gets removed as dust.

Weight: O(1) DB Weight:

Read: Bonded, Ledger, [Origin Account], Current Era, History Depth, Locks

Write: Bonded, Payee, [Origin Account], Locks, Ledger

NOTE: Calling this function will bypass origin filters.

pub fn validate(origin: T::Origin, prefs: ValidatorPrefs) -> DispatchResult

Declare the desire to validate for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when EraElectionStatus is Closed.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1) DB Weight:

Read: Era Election Status, Ledger

Write: Nominators, Validators

NOTE: Calling this function will bypass origin filters.

pub fn nominate(
origin: T::Origin,
targets: Vec<<T::Lookup as StaticLookup>::Source>
) -> DispatchResult

Declare the desire to nominate targets for the origin controller.

Effects will be felt at the beginning of the next era. This can only be called when EraElectionStatus is Closed.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when EraElectionStatus is Closed.

The transaction’s complexity is proportional to the size of targets (N) which is capped at CompactAssignments::LIMIT (MAX_NOMINATIONS).

Both the reads and writes follow a similar pattern.

Weight: O(N) where N is the number of targets DB Weight:

Reads: Era Election Status, Ledger, Current Era

Writes: Validators, Nominators

NOTE: Calling this function will bypass origin filters.

pub fn submit_election_solution(
origin: T::Origin,
winners: Vec<ValidatorIndex>,
compact: CompactAssignments,
score: ElectionScore,
era: EraIndex,
size: ElectionSize
) -> DispatchResultWithPostInfo

Submit an election result to the chain. If the solution:

is valid.

has a better score than a potentially existing solution on chain.

then, it will be put on chain.

A solution consists of two pieces of data:

winners: a flat vector of all the winners of the round.

assignments: the compact version of an assignment vector that encodes the edge weights.

Both of which may be computed using phragmen, or any other algorithm.

Additionally, the submitter must provide:

The score that they claim their solution has.

Both validators and nominators will be represented by indices in the solution. The indices should respect the corresponding types (ValidatorIndex and NominatorIndex). Moreover, they should be valid when used to index into SnapshotValidators and SnapshotNominators. Any invalid index will cause the solution to be rejected. These two storage items are set during the election window and may be used to determine the indices.

A solution is valid if:

It is submitted when EraElectionStatus is Open.

Its claimed score is equal to the score computed on-chain.

Presents the correct number of winners.

All indexes must be value according to the snapshot vectors. All edge values must also be correct and should not overflow the granularity of the ratio type (i.e. 256 or billion).

For each edge, all targets are actually nominated by the voter.

Has correct self-votes.

A solutions score is consisted of 3 parameters:

min { support.total } for each support of a winner. This value should be maximized.

sum { support.total } for each support of a winner. This value should be minimized.

sum { support.total^2 } for each support of a winner. This value should be minimized (to ensure less variance)

The transaction is assumed to be the longest path, a better solution.

Initial solution is almost the same.

Worse solution is retraced in pre-dispatch-checks which sets its own weight.

NOTE: Calling this function will bypass origin filters.

pub fn submit_election_solution_unsigned(
origin: T::Origin,
winners: Vec<ValidatorIndex>,
compact: CompactAssignments,
score: ElectionScore,
era: EraIndex,
size: ElectionSize
) -> DispatchResultWithPostInfo

Unsigned version of submit_election_solution.

Note that this must pass the [ValidateUnsigned] check which only allows transactions from the local node to be included. In other words, only the block author can include a transaction in the block.

See [submit_election_solution].

NOTE: Calling this function will bypass origin filters.

pub fn kick(
origin: T::Origin,
who: Vec<<T::Lookup as StaticLookup>::Source>
) -> DispatchResult

Remove the given nominations from the calling validator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when EraElectionStatus is Closed. The controller account should represent a validator.

who: A list of nominator stash accounts who are nominating this validator which should no longer be nominating this validator.

Note: Making this call only makes sense if you first set the validator preferences to block any further nominations.

NOTE: Calling this function will bypass origin filters.

impl<T: Config> Module<T>

pub fn slashable_balance_of(stash: &T::AccountId) -> BalanceOf<T>

The total balance that can be slashed from a stash account as of right now.

pub fn slashable_balance_of_vote_weight(
stash: &T::AccountId,
issuance: BalanceOf<T>
) -> VoteWeight

Internal impl of Self::slashable_balance_of that returns VoteWeight.

pub fn slashable_balance_of_fn() -> Box<dyn Fn(&T::AccountId) -> VoteWeight>

Returns a closure around slashable_balance_of_vote_weight that can be passed around.

This prevents call sites from repeatedly requesting total_issuance from backend. But it is important to be only used while the total issuance is not changing.

pub fn create_stakers_snapshot() -> (bool, Weight)

Dump the list of validators and nominators into vectors and keep them on-chain.

This data is used to efficiently evaluate election results. returns true if the operation is successful.

pub fn pre_dispatch_checks(
score: ElectionScore,
era: EraIndex
) -> DispatchResultWithPostInfo

Basic and cheap checks that we perform in validate unsigned, and in the execution.

State reads: ElectionState, CurrentEr, QueuedScore.

This function does weight refund in case of errors, which is based upon the fact that it is called at the very beginning of the call site’s function.

pub fn check_and_replace_solution(
winners: Vec<ValidatorIndex>,
compact_assignments: CompactAssignments,
compute: ElectionCompute,
claimed_score: ElectionScore,
era: EraIndex,
election_size: ElectionSize
) -> DispatchResultWithPostInfo

Checks a given solution and if correct and improved, writes it on chain as the queued result of the next round. This may be called by both a signed and an unsigned transaction.

pub fn do_on_chain_phragmen(
) -> Option<ElectionResult<T::AccountId, BalanceOf<T>>>

Execute election and return the new results. The edge weights are processed into support values.

This is basically a wrapper around Self::do_phragmen which translates PrimitiveElectionResult into ElectionResult.

No storage item is updated.

pub fn do_phragmen<Accuracy: PerThing128>(
iterations: usize
) -> Option<PrimitiveElectionResult<T::AccountId, Accuracy>>

Execute phragmen election and return the new results. No post-processing is applied and the raw edge weights are returned.

Self votes are added and nominations before the most recent slashing span are ignored.

No storage item is updated.

pub fn reward_by_ids(
validators_points: impl IntoIterator<Item = (T::AccountId, u32)>
)

Add reward points to validators using their stash account ID.

Validators are keyed by stash account ID and must be in the current elected set.

For each element in the iterator the given number of points in u32 is added to the validator, thus duplicates are handled.

At the end of the era each the total payout will be distributed among validator relatively to their points.

COMPLEXITY: Complexity is number_of_validator_to_reward x current_elected_len. If you need to reward lots of validator consider using reward_by_indices.