Struct pallet_staking::Module

pub struct Module<T: Config>(_);

Implementations

impl<T: Config + 'static> Module<T>

pub fn history_depth() -> u32

Number of eras to keep in history.

Information is kept for eras in [current_era - history_depth; current_era].

Must be more than the number of eras delayed by session otherwise. I.e. active era must always be in history. I.e. active_era > current_era - history_depth must be guaranteed.

pub fn validator_count() -> u32

The ideal number of staking participants.

pub fn minimum_validator_count() -> u32

Minimum number of staking participants before emergency conditions are imposed.

pub fn invulnerables() -> Vec<T::AccountId>

Any validators that may never be slashed or forcibly kicked. It's a Vec since they're easy to initialize and the performance hit is minimal (we expect no more than four invulnerables) and restricted to testnets.

pub fn bonded<K: EncodeLike<T::AccountId>>(key: K) -> Option<T::AccountId>

Map from all locked "stash" accounts to the controller account.

pub fn ledger<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<StakingLedger<T::AccountId, BalanceOf<T>>>

Map from all (unlocked) "controller" accounts to the info regarding the staking.

pub fn payee<K: EncodeLike<T::AccountId>>(
key: K
) -> RewardDestination<T::AccountId>

Where the reward payment should be made. Keyed by stash.

pub fn validators<K: EncodeLike<T::AccountId>>(key: K) -> ValidatorPrefs

The map from (wannabe) validator stash key to the preferences of that validator.

pub fn nominators<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<Nominations<T::AccountId>>

The map from nominator stash key to the set of stash keys of all validators to nominate.

pub fn current_era() -> Option<EraIndex>

The current era index.

This is the latest planned era, depending on how the Session pallet queues the validator set, it might be active or not.

pub fn active_era() -> Option<ActiveEraInfo>

The active era information, it holds index and start.

The active era is the era being currently rewarded. Validator set of this era must be equal to SessionInterface::validators.

pub fn eras_start_session_index<K: EncodeLike<EraIndex>>(
key: K
) -> Option<SessionIndex>

The session index at which the era start for the last HISTORY_DEPTH eras.

Note: This tracks the starting session (i.e. session index when era start being active) for the eras in [CurrentEra - HISTORY_DEPTH, CurrentEra].

pub fn eras_stakers<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Exposure<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

pub fn eras_stakers_clipped<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Exposure<T::AccountId, BalanceOf<T>> where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

pub fn eras_validator_prefs<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> ValidatorPrefs where
KArg1: EncodeLike<EraIndex>,
KArg2: EncodeLike<T::AccountId>,

pub fn eras_validator_reward<K: EncodeLike<EraIndex>>(
key: K
) -> Option<BalanceOf<T>>

The total validator era payout for the last HISTORY_DEPTH eras.

Eras that haven't finished yet or has been removed doesn't have reward.

pub fn eras_reward_points<K: EncodeLike<EraIndex>>(
key: K
) -> EraRewardPoints<T::AccountId>

Rewards for the last HISTORY_DEPTH eras. If reward hasn't been set or has been removed then 0 reward is returned.

pub fn eras_total_stake<K: EncodeLike<EraIndex>>(key: K) -> BalanceOf<T>

The total amount staked for the last HISTORY_DEPTH eras. If total hasn't been set or has been removed then 0 stake is returned.

pub fn force_era() -> Forcing

Mode of era forcing.

pub fn slash_reward_fraction() -> Perbill

The percentage of the slash that is distributed to reporters.

The rest of the slashed value is handled by the Slash.

pub fn canceled_payout() -> BalanceOf<T>

The amount of currency given to reporters of a slash event which was canceled by extraordinary circumstances (e.g. governance).

pub fn slashing_spans<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<SlashingSpans>

Slashing spans for stash accounts.

pub fn current_planned_session() -> SessionIndex

The last planned session scheduled by the session pallet.

This is basically in sync with the call to [SessionManager::new_session].

impl<T: Config> Module<T>

Can also be called using Call.

pub fn bond(
origin: T::Origin,
controller: <T::Lookup as StaticLookup>::Source,
value: BalanceOf<T>,
payee: RewardDestination<T::AccountId>
) -> DispatchResult

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

value must be more than the minimum_balance specified by T::Currency.

The dispatch origin for this call must be Signed by the stash account.

Emits Bonded.

Independent of the arguments. Moderate complexity.

O(1).

Three extra DB entries.

NOTE: Two of the storage writes (Self::bonded, Self::payee) are never cleaned unless the origin falls below existential deposit and gets removed as dust.

Weight: O(1) DB Weight:

Read: Bonded, Ledger, [Origin Account], Current Era, History Depth, Locks

Write: Bonded, Payee, [Origin Account], Locks, Ledger

NOTE: Calling this function will bypass origin filters.

pub fn validate(origin: T::Origin, prefs: ValidatorPrefs) -> DispatchResult

Declare the desire to validate for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

Weight: O(1) DB Weight:

Read: Era Election Status, Ledger

Write: Nominators, Validators

NOTE: Calling this function will bypass origin filters.

pub fn nominate(
origin: T::Origin,
targets: Vec<<T::Lookup as StaticLookup>::Source>
) -> DispatchResult

Declare the desire to nominate targets for the origin controller.

Effects will be felt at the beginning of the next era. This can only be called when [EraElectionStatus] is Closed.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed.

The transaction's complexity is proportional to the size of targets (N) which is capped at CompactAssignments::LIMIT (MAX_NOMINATIONS).

Both the reads and writes follow a similar pattern.

Weight: O(N) where N is the number of targets DB Weight:

Reads: Era Election Status, Ledger, Current Era

Writes: Validators, Nominators

NOTE: Calling this function will bypass origin filters.

pub fn kick(
origin: T::Origin,
who: Vec<<T::Lookup as StaticLookup>::Source>
) -> DispatchResult

Remove the given nominations from the calling validator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash. And, it can be only called when [EraElectionStatus] is Closed. The controller account should represent a validator.

who: A list of nominator stash accounts who are nominating this validator which should no longer be nominating this validator.

Note: Making this call only makes sense if you first set the validator preferences to block any further nominations.

NOTE: Calling this function will bypass origin filters.

impl<T: Config> Module<T>

pub fn slashable_balance_of(stash: &T::AccountId) -> BalanceOf<T>

The total balance that can be slashed from a stash account as of right now.

pub fn slashable_balance_of_vote_weight(
stash: &T::AccountId,
issuance: BalanceOf<T>
) -> VoteWeight

Internal impl of Self::slashable_balance_of that returns VoteWeight.

pub fn slashable_balance_of_fn() -> Box<dyn Fn(&T::AccountId) -> VoteWeight>

Returns a closure around slashable_balance_of_vote_weight that can be passed around.

This prevents call sites from repeatedly requesting total_issuance from backend. But it is important to be only used while the total issuance is not changing.

pub fn process_election(
flat_supports: Supports<T::AccountId>,
current_era: EraIndex
) -> Result<Vec<T::AccountId>, ()>

Process the output of the election.

This ensures enough validators have been elected, converts all supports to exposures and writes them to the associated storage.

Returns Err(()) if less than MinimumValidatorCount validators have been elected, Ok otherwise.

pub fn reward_by_ids(
validators_points: impl IntoIterator<Item = (T::AccountId, u32)>
)

Add reward points to validators using their stash account ID.

Validators are keyed by stash account ID and must be in the current elected set.

For each element in the iterator the given number of points in u32 is added to the validator, thus duplicates are handled.

At the end of the era each the total payout will be distributed among validator relatively to their points.

COMPLEXITY: Complexity is number_of_validator_to_reward x current_elected_len.

pub fn add_era_stakers(
current_era: EraIndex,
controller: T::AccountId,
exposure: Exposure<T::AccountId, BalanceOf<T>>
)

pub fn set_slash_reward_fraction(fraction: Perbill)

pub fn get_npos_voters() -> Vec<(T::AccountId, VoteWeight, Vec<T::AccountId>)>

Get all of the voters that are eligible for the npos election.

This will use all on-chain nominators, and all the validators will inject a self vote.

Slashing

All nominations that have been submitted before the last non-zero slash of the validator are auto-chilled.

Note that this is VERY expensive. Use with care.

pub fn get_npos_targets() -> Vec<T::AccountId>