Struct pallet_staking::Module

pub struct Module<T: Trait>(_);

Methods

impl<T: Trait + 'static> Module<T>

pub fn validator_count() -> u32

The ideal number of staking participants.

pub fn minimum_validator_count() -> u32

Minimum number of staking participants before emergency conditions are imposed.

pub fn invulnerables() -> Vec<T::AccountId>

Any validators that may never be slashed or forcibly kicked. It's a Vec since they're easy to initialize and the performance hit is minimal (we expect no more than four invulnerables) and restricted to testnets.

pub fn bonded<K: EncodeLike<T::AccountId>>(key: K) -> Option<T::AccountId>

Map from all locked "stash" accounts to the controller account.

pub fn ledger<K: EncodeLike<T::AccountId>>(
key: K
) -> Option<StakingLedger<T::AccountId, BalanceOf<T>>>

Map from all (unlocked) "controller" accounts to the info regarding the staking.

pub fn payee<K: EncodeLike<T::AccountId>>(key: K) -> RewardDestination

Where the reward payment should be made. Keyed by stash.

pub fn validators<K: EncodeLike<T::AccountId>>(
key: K
) -> ValidatorPrefs<BalanceOf<T>>

The map from (wannabe) validator stash key to the preferences of that validator.

pub fn nominators<K: EncodeLike<T::AccountId>>(key: K) -> Vec<T::AccountId>

The map from nominator stash key to the set of stash keys of all validators to nominate.

pub fn stakers<K: EncodeLike<T::AccountId>>(
key: K
) -> Exposure<T::AccountId, BalanceOf<T>>

Nominators for a particular account that is in action right now. You can't iterate through validators here, but you can find them in the Session module.

This is keyed by the stash account.

pub fn current_elected() -> Vec<T::AccountId>

The currently elected validator set keyed by stash account ID.

pub fn current_era() -> EraIndex

The current era index.

pub fn current_era_start() -> <<T as Trait>::Time as Time>::Moment

The start of the current era.

pub fn current_era_start_session_index() -> SessionIndex

The session index at which the current era started.

pub fn current_era_reward() -> EraPoints

Rewards for the current era. Using indices of current elected set.

pub fn slot_stake() -> BalanceOf<T>

The amount of balance actively at stake for each validator slot, currently.

This is used to derive rewards and punishments.

pub fn force_era() -> Forcing

True if the next session change will be a new era regardless of index.

pub fn slash_reward_fraction() -> Perbill

The percentage of the slash that is distributed to reporters.

The rest of the slashed value is handled by the Slash.

pub fn era_slash_journal<K: EncodeLike<EraIndex>>(
key: K
) -> Vec<SlashJournalEntry<T::AccountId, BalanceOf<T>>>

All slashes that have occurred in a given era.

impl<T: Trait> Module<T>

pub fn slashable_balance_of(stash: &T::AccountId) -> BalanceOf<T>

The total balance that can be slashed from a validator controller account as of right now.

pub fn reward_by_ids(
validators_points: impl IntoIterator<Item = (T::AccountId, u32)>
)

Add reward points to validators using their stash account ID.

Validators are keyed by stash account ID and must be in the current elected set.

For each element in the iterator the given number of points in u32 is added to the validator, thus duplicates are handled.

At the end of the era each the total payout will be distributed among validator relatively to their points.

COMPLEXITY: Complexity is number_of_validator_to_reward x current_elected_len. If you need to reward lots of validator consider using reward_by_indices.

pub fn reward_by_indices(
validators_points: impl IntoIterator<Item = (u32, u32)>
)

Add reward points to validators using their validator index.

For each element in the iterator the given number of points in u32 is added to the validator, thus duplicates are handled.