Struct pallet_staking::Pallet

pub struct Pallet<T>(_);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>

pub fn slashable_balance_of(stash: &T::AccountId) -> BalanceOf<T>

The total balance that can be slashed from a stash account as of right now.

pub fn slashable_balance_of_vote_weight( stash: &T::AccountId, issuance: BalanceOf<T> ) -> VoteWeight

Internal impl of Self::slashable_balance_of that returns VoteWeight.

pub fn weight_of_fn() -> Box<dyn Fn(&T::AccountId) -> VoteWeight>

Returns a closure around slashable_balance_of_vote_weight that can be passed around.

This prevents call sites from repeatedly requesting total_issuance from backend. But it is important to be only used while the total issuance is not changing.

pub fn weight_of(who: &T::AccountId) -> VoteWeight

Same as weight_of_fn, but made for one time use.

pub fn trigger_new_era( start_session_index: SessionIndex, exposures: BoundedVec<(T::AccountId, Exposure<T::AccountId, BalanceOf<T>>), MaxWinnersOf<T>> ) -> BoundedVec<T::AccountId, MaxWinnersOf<T>>

Plan a new era.

Bump the current era storage (which holds the latest planned era).

Store start session index for the new planned era.

Clean old era information.

Store staking information for the new planned era

Returns the new validator set.

pub fn store_stakers_info( exposures: BoundedVec<(T::AccountId, Exposure<T::AccountId, BalanceOf<T>>), MaxWinnersOf<T>>, new_planned_era: EraIndex ) -> BoundedVec<T::AccountId, MaxWinnersOf<T>>

Process the output of the election.

Store staking information for the new planned era

pub fn reward_by_ids( validators_points: impl IntoIterator<Item = (T::AccountId, u32)> )

Add reward points to validators using their stash account ID.

Validators are keyed by stash account ID and must be in the current elected set.

For each element in the iterator the given number of points in u32 is added to the validator, thus duplicates are handled.

At the end of the era each the total payout will be distributed among validator relatively to their points.

COMPLEXITY: Complexity is number_of_validator_to_reward x current_elected_len.

pub fn add_era_stakers( current_era: EraIndex, stash: T::AccountId, exposure: Exposure<T::AccountId, BalanceOf<T>> )

pub fn set_slash_reward_fraction(fraction: Perbill)

pub fn get_npos_voters(bounds: DataProviderBounds) -> Vec<VoterOf<Self>>

Get all of the voters that are eligible for the npos election.

maybe_max_len can imposes a cap on the number of voters returned;

Sets MinimumActiveStake to the minimum active nominator stake in the returned set of nominators.

This function is self-weighing as DispatchClass::Mandatory.

pub fn get_npos_targets(bounds: DataProviderBounds) -> Vec<T::AccountId>

Get the targets for an upcoming npos election.

This function is self-weighing as DispatchClass::Mandatory.

pub fn do_add_nominator(who: &T::AccountId, nominations: Nominations<T>)

This function will add a nominator to the Nominators storage map, and VoterList.

If the nominator already exists, their nominations will be updated.

NOTE: you must ALWAYS use this function to add nominator or update their targets. Any access to Nominators or VoterList outside of this function is almost certainly wrong.

pub fn do_remove_nominator(who: &T::AccountId) -> bool

This function will remove a nominator from the Nominators storage map, and VoterList.

Returns true if who was removed from Nominators, otherwise false.

NOTE: you must ALWAYS use this function to remove a nominator from the system. Any access to Nominators or VoterList outside of this function is almost certainly wrong.

pub fn do_add_validator(who: &T::AccountId, prefs: ValidatorPrefs)

This function will add a validator to the Validators storage map.

If the validator already exists, their preferences will be updated.

NOTE: you must ALWAYS use this function to add a validator to the system. Any access to Validators or VoterList outside of this function is almost certainly wrong.

pub fn do_remove_validator(who: &T::AccountId) -> bool

This function will remove a validator from the Validators storage map.

Returns true if who was removed from Validators, otherwise false.

NOTE: you must ALWAYS use this function to remove a validator from the system. Any access to Validators or VoterList outside of this function is almost certainly wrong.

impl<T: Config> Pallet<T>

pub fn api_nominations_quota(balance: BalanceOf<T>) -> u32

Returns the current nominations quota for nominators.

Used by the runtime API.

impl<T: Config> Pallet<T>

pub fn bond( origin: OriginFor<T>, value: BalanceOf<T>, payee: RewardDestination<T::AccountId> ) -> DispatchResult

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

value must be more than the minimum_balance specified by T::Currency.

The dispatch origin for this call must be Signed by the stash account.

Emits Bonded.

Complexity

Independent of the arguments. Moderate complexity.

O(1).

Three extra DB entries.

NOTE: Two of the storage writes (Self::bonded, Self::payee) are never cleaned unless the origin falls below existential deposit and gets removed as dust.

pub fn bond_extra( origin: OriginFor<T>, max_additional: BalanceOf<T> ) -> DispatchResult

Add some extra amount that have appeared in the stash free_balance into the balance up for staking.

The dispatch origin for this call must be Signed by the stash, not the controller.

Use this if there are additional funds in your stash account that you wish to bond. Unlike bond or unbond this function does not impose any limitation on the amount that can be added.

Emits Bonded.

Complexity

Independent of the arguments. Insignificant complexity.

O(1).

pub fn unbond( origin: OriginFor<T>, value: BalanceOf<T> ) -> DispatchResultWithPostInfo

Schedule a portion of the stash to be unlocked ready for transfer out after the bond period ends. If this leaves an amount actively bonded less than T::Currency::minimum_balance(), then it is increased to the full amount.

The dispatch origin for this call must be Signed by the controller, not the stash.

Once the unlock period is done, you can call withdraw_unbonded to actually move the funds out of management ready for transfer.

No more than a limited number of unlocking chunks (see MaxUnlockingChunks) can co-exists at the same time. If there are no unlocking chunks slots available Call::withdraw_unbonded is called to remove some of the chunks (if possible).

If a user encounters the InsufficientBond error when calling this extrinsic, they should call chill first in order to free up their bonded funds.

Emits Unbonded.

See also Call::withdraw_unbonded.

pub fn withdraw_unbonded( origin: OriginFor<T>, num_slashing_spans: u32 ) -> DispatchResultWithPostInfo

Remove any unlocked chunks from the unlocking queue from our management.

This essentially frees up that balance to be used by the stash account to do whatever it wants.

The dispatch origin for this call must be Signed by the controller.

Emits Withdrawn.

See also Call::unbond.

Parameters

num_slashing_spans indicates the number of metadata slashing spans to clear when this call results in a complete removal of all the data related to the stash account. In this case, the num_slashing_spans must be larger or equal to the number of slashing spans associated with the stash account in the SlashingSpans storage type, otherwise the call will fail. The call weight is directly propotional to num_slashing_spans.

Complexity

O(S) where S is the number of slashing spans to remove NOTE: Weight annotation is the kill scenario, we refund otherwise.

pub fn validate(origin: OriginFor<T>, prefs: ValidatorPrefs) -> DispatchResult

Declare the desire to validate for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

pub fn nominate( origin: OriginFor<T>, targets: Vec<<<T as Config>::Lookup as StaticLookup>::Source> ) -> DispatchResult

Declare the desire to nominate targets for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

Complexity

The transaction’s complexity is proportional to the size of targets (N) which is capped at CompactAssignments::LIMIT (T::MaxNominations).

Both the reads and writes follow a similar pattern.

pub fn chill(origin: OriginFor<T>) -> DispatchResult

Declare no desire to either validate or nominate.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

Complexity

Independent of the arguments. Insignificant complexity.

Contains one read.

Writes are limited to the origin account key.

pub fn set_payee( origin: OriginFor<T>, payee: RewardDestination<T::AccountId> ) -> DispatchResult

(Re-)set the payment target for a controller.

Effects will be felt instantly (as soon as this function is completed successfully).

The dispatch origin for this call must be Signed by the controller, not the stash.

Complexity

O(1)

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

pub fn set_controller(origin: OriginFor<T>) -> DispatchResult

(Re-)sets the controller of a stash to the stash itself. This function previously accepted a controller argument to set the controller to an account other than the stash itself. This functionality has now been removed, now only setting the controller to the stash, if it is not already.

Effects will be felt instantly (as soon as this function is completed successfully).

The dispatch origin for this call must be Signed by the stash, not the controller.

Complexity

O(1)

Independent of the arguments. Insignificant complexity.

Contains a limited number of reads.

Writes are limited to the origin account key.

pub fn set_validator_count(origin: OriginFor<T>, new: u32) -> DispatchResult

Sets the ideal number of validators.

The dispatch origin must be Root.

Complexity

O(1)

pub fn increase_validator_count( origin: OriginFor<T>, additional: u32 ) -> DispatchResult

Increments the ideal number of validators upto maximum of ElectionProviderBase::MaxWinners.

The dispatch origin must be Root.

Complexity

Same as Self::set_validator_count.

pub fn scale_validator_count( origin: OriginFor<T>, factor: Percent ) -> DispatchResult

Scale up the ideal number of validators by a factor upto maximum of ElectionProviderBase::MaxWinners.

The dispatch origin must be Root.

Complexity

Same as Self::set_validator_count.

pub fn force_no_eras(origin: OriginFor<T>) -> DispatchResult

Force there to be no new eras indefinitely.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. Thus the election process may be ongoing when this is called. In this case the election will continue until the next era is triggered.

Complexity

No arguments.

Weight: O(1)

pub fn force_new_era(origin: OriginFor<T>) -> DispatchResult

Force there to be a new era at the end of the next session. After this, it will be reset to normal (non-forced) behaviour.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. If this is called just before a new era is triggered, the election process may not have enough blocks to get a result.

Complexity

No arguments.

Weight: O(1)

pub fn set_invulnerables( origin: OriginFor<T>, invulnerables: Vec<T::AccountId> ) -> DispatchResult

Set the validators who cannot be slashed (if any).

The dispatch origin must be Root.

pub fn force_unstake( origin: OriginFor<T>, stash: T::AccountId, num_slashing_spans: u32 ) -> DispatchResult

Force a current staker to become completely unstaked, immediately.

The dispatch origin must be Root.

Parameters

num_slashing_spans: Refer to comments on Call::withdraw_unbonded for more details.

pub fn force_new_era_always(origin: OriginFor<T>) -> DispatchResult

Force there to be a new era at the end of sessions indefinitely.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. If this is called just before a new era is triggered, the election process may not have enough blocks to get a result.

pub fn cancel_deferred_slash( origin: OriginFor<T>, era: EraIndex, slash_indices: Vec<u32> ) -> DispatchResult

Cancel enactment of a deferred slash.

Can be called by the T::AdminOrigin.

Parameters: era and indices of the slashes for that era to kill.

pub fn payout_stakers( origin: OriginFor<T>, validator_stash: T::AccountId, era: EraIndex ) -> DispatchResultWithPostInfo

Pay out all the stakers behind a single validator for a single era.

validator_stash is the stash account of the validator. Their nominators, up to T::MaxNominatorRewardedPerValidator, will also receive their rewards.

era may be any era between [current_era - history_depth; current_era].

The origin of this call must be Signed. Any account can call this function, even if it is not one of the stakers.

Complexity

At most O(MaxNominatorRewardedPerValidator).

pub fn rebond( origin: OriginFor<T>, value: BalanceOf<T> ) -> DispatchResultWithPostInfo

Rebond a portion of the stash scheduled to be unlocked.

The dispatch origin must be signed by the controller.

Complexity

Time complexity: O(L), where L is unlocking chunks

Bounded by MaxUnlockingChunks.

pub fn reap_stash( origin: OriginFor<T>, stash: T::AccountId, num_slashing_spans: u32 ) -> DispatchResultWithPostInfo

Remove all data structures concerning a staker/stash once it is at a state where it can be considered dust in the staking system. The requirements are:

the total_balance of the stash is below existential deposit.

or, the ledger.total of the stash is below existential deposit.

The former can happen in cases like a slash; the latter when a fully unbonded account is still receiving staking rewards in RewardDestination::Staked.

It can be called by anyone, as long as stash meets the above requirements.

Refunds the transaction fees upon successful execution.

Parameters

num_slashing_spans: Refer to comments on Call::withdraw_unbonded for more details.

pub fn kick( origin: OriginFor<T>, who: Vec<<<T as Config>::Lookup as StaticLookup>::Source> ) -> DispatchResult

Remove the given nominations from the calling validator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

who: A list of nominator stash accounts who are nominating this validator which should no longer be nominating this validator.

Note: Making this call only makes sense if you first set the validator preferences to block any further nominations.

pub fn set_staking_configs( origin: OriginFor<T>, min_nominator_bond: ConfigOp<BalanceOf<T>>, min_validator_bond: ConfigOp<BalanceOf<T>>, max_nominator_count: ConfigOp<u32>, max_validator_count: ConfigOp<u32>, chill_threshold: ConfigOp<Percent>, min_commission: ConfigOp<Perbill> ) -> DispatchResult

Update the various staking configurations .

min_nominator_bond: The minimum active bond needed to be a nominator.

min_validator_bond: The minimum active bond needed to be a validator.

max_nominator_count: The max number of users who can be a nominator at once. When set to None, no limit is enforced.

max_validator_count: The max number of users who can be a validator at once. When set to None, no limit is enforced.

chill_threshold: The ratio of max_nominator_count or max_validator_count which should be filled in order for the chill_other transaction to work.

min_commission: The minimum amount of commission that each validators must maintain. This is checked only upon calling validate. Existing validators are not affected.

RuntimeOrigin must be Root to call this function.

NOTE: Existing nominators and validators will not be affected by this update. to kick people under the new limits, chill_other should be called.

pub fn chill_other( origin: OriginFor<T>, controller: T::AccountId ) -> DispatchResult

Declare a controller to stop participating as either a validator or nominator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed, but can be called by anyone.

If the caller is the same as the controller being targeted, then no further checks are enforced, and this function behaves just like chill.

If the caller is different than the controller being targeted, the following conditions must be met:

controller must belong to a nominator who has become non-decodable,

Or:

A ChillThreshold must be set and checked which defines how close to the max nominators or validators we must reach before users can start chilling one-another.

A MaxNominatorCount and MaxValidatorCount must be set which is used to determine how close we are to the threshold.

A MinNominatorBond and MinValidatorBond must be set and checked, which determines if this is a person that should be chilled because they have not met the threshold bond required.

This can be helpful if bond requirements are updated, and we need to remove old users who do not satisfy these requirements.

pub fn force_apply_min_commission( origin: OriginFor<T>, validator_stash: T::AccountId ) -> DispatchResult

Force a validator to have at least the minimum commission. This will not affect a validator who already has a commission greater than or equal to the minimum. Any account can call this.

pub fn set_min_commission(origin: OriginFor<T>, new: Perbill) -> DispatchResult

Sets the minimum amount of commission that each validators must maintain.

This call has lower privilege requirements than set_staking_config and can be called by the T::AdminOrigin. Root can always call this.

impl<T: Config> Pallet<T>

pub fn validator_count() -> u32

An auto-generated getter for ValidatorCount.

impl<T: Config> Pallet<T>

pub fn minimum_validator_count() -> u32

An auto-generated getter for MinimumValidatorCount.

impl<T: Config> Pallet<T>

pub fn invulnerables() -> Vec<T::AccountId>

An auto-generated getter for Invulnerables.

impl<T: Config> Pallet<T>

pub fn bonded<KArg>(k: KArg) -> Option<T::AccountId>where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for Bonded.

impl<T: Config> Pallet<T>

pub fn ledger<KArg>(k: KArg) -> Option<StakingLedger<T>>where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for Ledger.

impl<T: Config> Pallet<T>

pub fn payee<KArg>(k: KArg) -> RewardDestination<T::AccountId>where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for Payee.

impl<T: Config> Pallet<T>

pub fn validators<KArg>(k: KArg) -> ValidatorPrefswhere KArg: EncodeLike<T::AccountId>,

An auto-generated getter for Validators.

impl<T: Config> Pallet<T>

pub fn nominators<KArg>(k: KArg) -> Option<Nominations<T>>where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for Nominators.

impl<T: Config> Pallet<T>

pub fn current_era() -> Option<EraIndex>

An auto-generated getter for CurrentEra.

impl<T: Config> Pallet<T>

pub fn active_era() -> Option<ActiveEraInfo>

An auto-generated getter for ActiveEra.

impl<T: Config> Pallet<T>

pub fn eras_start_session_index<KArg>(k: KArg) -> Option<SessionIndex>where KArg: EncodeLike<EraIndex>,

An auto-generated getter for ErasStartSessionIndex.

impl<T: Config> Pallet<T>

pub fn eras_stakers<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Exposure<T::AccountId, BalanceOf<T>>where KArg1: EncodeLike<EraIndex>, KArg2: EncodeLike<T::AccountId>,

An auto-generated getter for ErasStakers.

impl<T: Config> Pallet<T>

pub fn eras_stakers_clipped<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Exposure<T::AccountId, BalanceOf<T>>where KArg1: EncodeLike<EraIndex>, KArg2: EncodeLike<T::AccountId>,

An auto-generated getter for ErasStakersClipped.

impl<T: Config> Pallet<T>

pub fn eras_validator_prefs<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> ValidatorPrefswhere KArg1: EncodeLike<EraIndex>, KArg2: EncodeLike<T::AccountId>,

An auto-generated getter for ErasValidatorPrefs.

impl<T: Config> Pallet<T>

pub fn eras_validator_reward<KArg>(k: KArg) -> Option<BalanceOf<T>>where KArg: EncodeLike<EraIndex>,

An auto-generated getter for ErasValidatorReward.

impl<T: Config> Pallet<T>

pub fn eras_reward_points<KArg>(k: KArg) -> EraRewardPoints<T::AccountId>where KArg: EncodeLike<EraIndex>,

An auto-generated getter for ErasRewardPoints.

impl<T: Config> Pallet<T>

pub fn eras_total_stake<KArg>(k: KArg) -> BalanceOf<T>where KArg: EncodeLike<EraIndex>,

An auto-generated getter for ErasTotalStake.

impl<T: Config> Pallet<T>

pub fn force_era() -> Forcing

An auto-generated getter for ForceEra.

impl<T: Config> Pallet<T>

pub fn slash_reward_fraction() -> Perbill

An auto-generated getter for SlashRewardFraction.

impl<T: Config> Pallet<T>

pub fn canceled_payout() -> BalanceOf<T>

An auto-generated getter for CanceledSlashPayout.

impl<T: Config> Pallet<T>

pub fn slashing_spans<KArg>(k: KArg) -> Option<SlashingSpans>where KArg: EncodeLike<T::AccountId>,

An auto-generated getter for SlashingSpans.

impl<T: Config> Pallet<T>

pub fn current_planned_session() -> SessionIndex

An auto-generated getter for CurrentPlannedSession.

impl<T: Config> Pallet<T>

pub fn offending_validators() -> Vec<(u32, bool)>

An auto-generated getter for OffendingValidators.