Module pallet_staking::testing_utils

Testing utils for staking. Provides some common functions to setup staking state, such as bonding validators, nominators, and generating different types of solutions.

Functions

create_assignments_for_offchain
Create random assignments for the given list of winners. Each assignment will have MAX_NOMINATIONS edges.

create_funded_user
Grab a funded user.

create_stash_and_dead_controller
Create a stash and controller pair, where the controller is dead, and payouts go to controller. This is used to test worst case payout scenarios.

create_stash_controller
Create a stash and controller pair.

create_validators
create max validators.

create_validators_with_nominators_for_era
This function generates validators and nominators who are randomly nominating edge_per_nominator random validators (until to_nominate if provided).

current_era
get the active era.

get_seq_phragmen_solution
Create a solution for seq-phragmen. This uses the same internal function as used by the offchain worker code.

get_single_winner_solution
Returns a solution in which only one winner is elected with just a self vote.

get_weak_solution
Build a really bad but acceptable solution for election. This should always yield a solution which has a less score than the seq-phragmen.

init_active_era
initialize the first era.