Module pallet_staking::testing_utils

Testing utils for staking. Needs the testing-utils feature to be enabled.

Note that these helpers should NOT be used with the actual crate tests, but are rather designed for when the module is being externally tested (i.e. fuzzing, benchmarking, e2e tests). Enabling this feature in the current crate's Cargo.toml will leak all of this into a normal release build. Just don't do it.

Constants

USER
A dummy suer.

Functions

account
Build an account with the given index.

active_era
get the active era.

address
Build an address given Index

bond_nominator

bond_validator
Bond a validator.

clean
Remove all validator, nominators, votes and exposures.

get_seq_phragmen_solution
Create a solution for seq-phragmen. This uses the same internal function as used by the offchain worker code.

get_weak_solution
Build a really bad but acceptable solution for election. This should always yield a solution which has a less score than the seq-phragmen.

init_active_era
initialize the first era.

random
Random number in the range [a, b].

set_validator_count
Set the desired validator count, with related storage items.

setup_chain_stakers
Bond nun_validators validators and num_nominator nominators with edge_per_voter random votes per nominator.

signed
Generate signed origin from who.

signed_account
Generate signed origin from index.

Type Definitions

AddressOf
Address type of the T