Trait pallet_staking::Config

pub trait Config: Config + SendTransactionTypes<Call<Self>> {

    type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>;
    type UnixTime: UnixTime;
    type CurrencyToVote: CurrencyToVote<BalanceOf<Self>>;
    type RewardRemainder: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>;
    type Event: From<Event<Self>> + Into<<Self as Config>::Event>;
    type Slash: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>;
    type Reward: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::PositiveImbalance>;
    type SessionsPerEra: Get<SessionIndex>;
    type BondingDuration: Get<EraIndex>;
    type SlashDeferDuration: Get<EraIndex>;
    type SlashCancelOrigin: EnsureOrigin<Self::Origin>;
    type SessionInterface: SessionInterface<Self::AccountId>;
    type RewardCurve: Get<&'static PiecewiseLinear<'static>>;
    type NextNewSession: EstimateNextNewSession<Self::BlockNumber>;
    type ElectionLookahead: Get<Self::BlockNumber>;
    type Call: Dispatchable + From<Call<Self>> + IsSubType<Call<Self>> + Clone;
    type MaxIterations: Get<u32>;
    type MinSolutionScoreBump: Get<Perbill>;
    type MaxNominatorRewardedPerValidator: Get<u32>;
    type UnsignedPriority: Get<TransactionPriority>;
    type OffchainSolutionWeightLimit: Get<Weight>;
    type WeightInfo: WeightInfo;

}

Associated Types

type Currency: LockableCurrency<Self::AccountId, Moment = Self::BlockNumber>

The staking balance.

type UnixTime: UnixTime

Time used for computing era duration.

It is guaranteed to start being called from the first on_finalize. Thus value at genesis is not used.

type CurrencyToVote: CurrencyToVote<BalanceOf<Self>>

Convert a balance into a number used for election calculation. This must fit into a u64 but is allowed to be sensibly lossy. The u64 is used to communicate with the sp_npos_elections crate which accepts u64 numbers and does operations in 128. Consequently, the backward convert is used convert the u128s from sp-elections back to a BalanceOf.

type RewardRemainder: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>

Tokens have been minted and are unused for validator-reward. See Era payout.

type Event: From<Event<Self>> + Into<<Self as Config>::Event>

The overarching event type.

type Slash: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>

Handler for the unbalanced reduction when slashing a staker.

type Reward: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::PositiveImbalance>

Handler for the unbalanced increment when rewarding a staker.

type SessionsPerEra: Get<SessionIndex>

Number of sessions per era.

type BondingDuration: Get<EraIndex>

Number of eras that staked funds must remain bonded for.

type SlashDeferDuration: Get<EraIndex>

Number of eras that slashes are deferred by, after computation.

This should be less than the bonding duration. Set to 0 if slashes should be applied immediately, without opportunity for intervention.

type SlashCancelOrigin: EnsureOrigin<Self::Origin>

The origin which can cancel a deferred slash. Root can always do this.

type SessionInterface: SessionInterface<Self::AccountId>

Interface for interacting with a session module.

type RewardCurve: Get<&'static PiecewiseLinear<'static>>

The NPoS reward curve used to define yearly inflation. See Era payout.

type NextNewSession: EstimateNextNewSession<Self::BlockNumber>

Something that can estimate the next session change, accurately or as a best effort guess.

type ElectionLookahead: Get<Self::BlockNumber>

The number of blocks before the end of the era from which election submissions are allowed.

Setting this to zero will disable the offchain compute and only on-chain seq-phragmen will be used.

This is bounded by being within the last session. Hence, setting it to a value more than the length of a session will be pointless.

type Call: Dispatchable + From<Call<Self>> + IsSubType<Call<Self>> + Clone

The overarching call type.

type MaxIterations: Get<u32>

Maximum number of balancing iterations to run in the offchain submission.

If set to 0, balance_solution will not be executed at all.

type MinSolutionScoreBump: Get<Perbill>

The threshold of improvement that should be provided for a new solution to be accepted.

type MaxNominatorRewardedPerValidator: Get<u32>

The maximum number of nominators rewarded for each validator.

For each validator only the $MaxNominatorRewardedPerValidator biggest stakers can claim their reward. This used to limit the i/o cost for the nominator payout.

type UnsignedPriority: Get<TransactionPriority>

A configuration for base priority of unsigned transactions.

This is exposed so that it can be tuned for particular runtime, when multiple pallets send unsigned transactions.

type OffchainSolutionWeightLimit: Get<Weight>

Maximum weight that the unsigned transaction can have.

Chose this value with care. On one hand, it should be as high as possible, so the solution can contain as many nominators/validators as possible. On the other hand, it should be small enough to fit in the block.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.