Trait pallet_staking::SessionInterface

pub trait SessionInterface<AccountId> {
    // Required methods
    fn disable_validator(validator_index: u32) -> bool;
    fn validators() -> Vec<AccountId>;
    fn prune_historical_up_to(up_to: SessionIndex);
}

Means for interacting with a specialized version of the session trait.

This is needed because Staking sets the ValidatorIdOf of the pallet_session::Config

Required Methods

fn disable_validator(validator_index: u32) -> bool

Disable the validator at the given index, returns false if the validator was already disabled or the index is out of bounds.

fn validators() -> Vec<AccountId>

Get the validators from session.

fn prune_historical_up_to(up_to: SessionIndex)

Prune historical session tries up to but not including the given index.